% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_admin_labels}
\alias{add_admin_labels}
\title{Wrapper around `tmap::tm_text()` with sane defaults for plotting admin labels.}
\usage{
add_admin_labels(
  point,
  text,
  size = 0.5,
  fontface = "bold",
  fontfamily = "Leelawadee",
  shadow = TRUE,
  auto_placement = FALSE,
  remove_overlap = FALSE,
  ...
)
}
\arguments{
\item{point}{Multipoint shape defined by sf package.}

\item{text}{Text labels column.}

\item{size}{Relative size of the text labels.}

\item{fontface}{Fontface.}

\item{fontfamily}{Fontfamily. Leelawadee is your precious.}

\item{shadow}{Boolean. Add a shadow around text labels. Issue opened on Github to request.}

\item{auto_placement}{Logical that determines whether the labels are placed automatically.}

\item{remove_overlap}{Logical that determines whether the overlapping labels are removed.}

\item{...}{Other arguments to pass to `tmap::tm_text()`.}
}
\value{
A tmap layer.
}
\description{
Wrapper around `tmap::tm_text()` with sane defaults for plotting admin labels.
}
