% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_compass}
\alias{add_compass}
\title{Add a compass}
\usage{
add_compass(
  text_size = 0.6,
  position = c("right", 0.8),
  color_dark = cols_reach("black"),
  text_color = cols_reach("black"),
  type = "4star",
  ...
)
}
\arguments{
\item{text_size}{Relative font size.}

\item{position}{Position of the compass. Vector of two values, specifying the x and y coordinates.}

\item{color_dark}{Color of the dark parts of the compass.}

\item{text_color}{color of the text.}

\item{type}{Compass type, one of: "arrow", "4star", "8star", "radar", "rose".}

\item{...}{Other arguments to pass to `tmap::tm_compass()`.}
}
\value{
A tmap layer.
}
\description{
Add a compass
}
