% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_credits}
\alias{add_credits}
\title{Do you want to credit someone or some institution?}
\usage{
add_credits(text, size = 0.4, bg_color = NA, position = c(0.75, 0.02), ...)
}
\arguments{
\item{text}{Text.}

\item{size}{Relative text size.}

\item{bg_color}{Background color.}

\item{position}{Position. Vector of two coordinates. Usually somewhere down.}

\item{...}{Other arguments to pass to `tmap::tm_credits()`.}
}
\value{
A tmap layer.
}
\description{
Do you want to credit someone or some institution?
}
