% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_layout}
\alias{add_layout}
\title{Basic defaults based on `tmap::tm_layout()`}
\usage{
add_layout(
  title = NULL,
  legend_position = c(0.02, 0.5),
  frame = FALSE,
  legend_frame = cols_reach("main_grey"),
  legend_text_size = 0.6,
  legend_title_size = 0.8,
  title_size = 0.9,
  title_fontface = "bold",
  title_color = cols_reach("main_grey"),
  fontfamily = "Leelawadee",
  ...
)
}
\arguments{
\item{title}{Map title.}

\item{legend_position}{Legend position. Not above the map is a good start.}

\item{frame}{Boolean. Legend frame?}

\item{legend_frame}{Legend frame color.}

\item{legend_text_size}{Legend text size in 'pt'.}

\item{legend_title_size}{Legend title size in 'pt'.}

\item{title_size}{Title text size in 'pt'.}

\item{title_fontface}{Title fontface. Bold if you wanna exemplify a lot what it is about.}

\item{title_color}{Title font color.}

\item{fontfamily}{Overall fontfamily. Leelawadee is your precious.}

\item{...}{Other arguments to pass to `tmap::tm_layout()`.}
}
\value{
A tmap layer.
}
\description{
Basic defaults based on `tmap::tm_layout()`
}
