% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_scale_bar}
\alias{add_scale_bar}
\title{Add a scale bar}
\usage{
add_scale_bar(
  text_size = 0.6,
  position = c("left", 0.01),
  color_dark = cols_reach("black"),
  breaks = c(0, 50, 100),
  ...
)
}
\arguments{
\item{text_size}{Relative font size.}

\item{position}{Position of the compass. Vector of two values, specifying the x and y coordinates.}

\item{color_dark}{Color of the dark parts of the compass.}

\item{breaks}{Breaks of the scale bar. If not specified, breaks will be automatically be chosen given the prefered width of the scale bar. Example: c(0, 50, 100).}

\item{...}{Other arguments to pass to `tmap::tm_compass()`.}
}
\value{
A tmap layer.
}
\description{
Add a scale bar
}
