% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addModules.R
\name{addModule}
\alias{addModule}
\title{Add modules to view list}
\usage{
addModule(view, type)
}
\arguments{
\item{view}{the view list object to be modified}

\item{type}{the type of module to be added. The types currently supported are spectra_displayer.}
}
\value{
modified view object list
}
\description{
This function adds module to a view list object
}
\examples{
v <- createView( name = "visualizeR" )
v <- addModule( view = v, type="spectra_displayer")
}
