% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\name{aggFun}
\alias{aggFun}
\title{Aggregating}
\usage{
aggFun(praemie, category, data, decreasing = F)
}
\arguments{
\item{praemie}{}

\item{category}{}

\item{data}{}

\item{decreasing}{}
}
\description{
\code{aggFun}
}

