% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Define the Server Logic for the Shiny App}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny parameters}
}
\value{
None. The function is used for its side effects.
}
\description{
Define the Server Logic for the Shiny App
}
\keyword{internal}
