% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{app_ui}
\alias{app_ui}
\title{Define the User Interface for the Shiny App}
\usage{
app_ui()
}
\value{
A Shiny UI definition using \code{fluidPage()}.
}
\description{
Define the User Interface for the Shiny App
}
\keyword{internal}
