% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendData.R
\name{appendData}
\alias{appendData}
\title{This function allows to create, append data to be passed to the https://github.com/npellet/visualizer.
Prior to this, the list have to be parsed and stored as a json file. This file can be either stored locally
or stored on https://gist.github.com.}
\usage{
appendData(data, variable, variableName, type)
}
\arguments{
\item{data}{data, if it exists}

\item{variable}{new dataset}

\item{variableName}{name of the dataset. The name of the data must be exposed to the https://github.com/npellet/visualizer
using the exposeData.R function.}

\item{type}{type of the data. The types currently supported are, "1D Y array", "chart", "multiChart".}
}
\value{
a list object with the data
}
\description{
This function allows to create, append data to be passed to the https://github.com/npellet/visualizer.
Prior to this, the list have to be parsed and stored as a json file. This file can be either stored locally
or stored on https://gist.github.com.
}
\examples{
x <- seq(from = 0, to = pi, by = 0.1)

d <- appendData(variableName = "simpleArray", variable = cos(x), type = "1D Y array")
d <- appendData(data = d, variableName = "simpleArray2", variable = sin(x), type = "1D Y array")
toJSON(d)

chart <- data.frame("x" = x, "y" = cos(3*x), "highlight" = seq_along(x) - 1, "info"= paste0("ID: ", 0:31))
d <- appendData(data = d, variableName = "chart", variable = chart, type = "chart")
}
\seealso{
exposeData.R
}
