% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR5regions}
\alias{AR5regions}
\title{AR5 regions from the IPCC 5th Assessment Report.}
\format{
A \code{\link[sp]{SpatialPolygons-class}} object
}
\source{
Shapefile imported with function \code{\link[rgdal]{readOGR}} (\url{http://www.ipcc-data.org/guidelines/pages/ar5_regions.html})
}
\description{
Spatial Polygons of the geographical regions used in the IPCC 5th Assessment Report.
}
\details{
For further detail, see \url{http://www.ipcc-data.org/guidelines/pages/ar5_regions.html}
}
\examples{
\dontrun{
data(AR5regions)
require(sp)
require(transformeR)
sp::plot(AR5regions, border = "red", axes = TRUE)
draw.world.lines()
title("AR5 World regions")
}
}
