% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR6_WGI_referenceRegions}
\alias{AR6_WGI_referenceRegions}
\title{AR6 WGI regions from the IPCC 6th Assessment Report (land and ocean).}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the IPCC 6th Assessment Report.
}
\details{
Display region names : names(AR6_WGI_referenceRegions) or names(names(AR6_WGI_referenceRegions)) 
to only obtain the long names.
}
\examples{
\dontrun{
data(AR6_WGI_referenceRegions)

# All regions
sp::plot(AR6_WGI_referenceRegions, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 World land regions")

# A single region (use region names for subsetting)
sp::plot(AR6_WGI_referenceRegions["MED"], border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 Mediterranean region")
}
}
