% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR6regions_land}
\alias{AR6regions_land}
\title{AR6 land regions from the IPCC 6th Assessment Report.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the IPCC 6th Assessment Report.
}
\details{
Region names are:  "ALA",  "NEC",  "GIC",  "NWN",  "SWN",  "CNA",  "ENA",  "CAM",  "CAR",  "NWS",  
"SAM",  "SSA",  "SWS",  "SES",  "AMZ",  "NEB",  "NEU",  "CEU", "MED",  "SAH",  "WAF",  "NEAF", "CEAF", "SWAF", 
"SEAF", "CAF",  "NEA",  "NWA",  "WAS",  "CAS",  "TIB",  "EAS",  "SAS",  "SEA",  "NAU",  "SAU", "ANT",  "ARCO".
}
\examples{
\dontrun{
data(AR6regions_land)

# All regions
sp::plot(AR6regions_land, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 World land regions")

# A single region
sp::plot(AR6regions_land["MED"], border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 Mediterranean region")
}
}
