% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR6regions_ocean}
\alias{AR6regions_ocean}
\title{AR6 oceanic regions from the IPCC 6th Assessment Report.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the IPCC 6th Assessment Report.
}
\details{
Continent names are: "ARCO",  "SOO",  "SAO",  "EAO",  "NAO",  "EIO",  "SIO",  "ARS",  "BOB",  "SPO",  "EPO",  "NPO", "SPO*", "EPO*", "NPO*".
}
\examples{
\dontrun{
data(AR6regions_ocean)

#All continents (all regions)
sp::plot(AR6regions_ocean, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 oceanic regions")

#A single region
sp::plot(AR6regions_ocean["SOO"], border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 SOO")
}
}
