% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.MrGrid.R
\name{as.MrGrid}
\alias{as.MrGrid}
\title{Convert a gridded data set with just one ensemble member to a S4 class.}
\usage{
as.MrGrid(obj.s3)
}
\arguments{
\item{obj.s3}{A list the with the elements Variable, Dates, Data, xyCoords, InitializationDates 
and Members}
}
\description{
Convert a gridded data set with just one ensemble member to the MrGrid class. 
 MrGrid is a S4 class derived from the MrData. The data set must have the elements Variable, 
 Dates, Data, xyCoords, InitializationDates and Members. This function is prepared to convert 
 the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See the loadeR.ECOMS R 
 package for more details (http://github.com/SantanderMetGroup/loadeR.ECOMS).
}
\details{
The visualization functions are programmed to work with S4 class. This function converts gridded 
 data sets (with just one member) to S4 to use those functions.
}
\note{
For gridded data sets with more than one ensemble member use as.MrEnsemble function.
 For station data use as.MrStation function.
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es}, J. Fernandez and Max Tuni
}
\seealso{
Other VisualizeR: \code{\link{as.MrEnsemble}},
  \code{\link{as.MrStation}}, \code{\link{bubblePlotS4}},
  \code{\link{convertIntoS4}}
}

