% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.MrStation.R
\name{as.MrStation}
\alias{as.MrStation}
\title{Convert station data to a S4 class.}
\usage{
as.MrStation(obj.s3)
}
\arguments{
\item{obj.s3}{A list the with the elements Variable, Dates, Data, xyCoords and MetaData.}
}
\description{
Convert station data to the MrStation class. MrStation is a S4 class derived 
 from the MrData. The data set must have the elements Variable, Dates, Data, xyCoords and 
 MetaData. This function is prepared to convert the station data loaded from the ECOMS User 
 Data Gateway (ECOMS-UDG). See the loadeR.ECOMS R package for more details 
 (http://github.com/SantanderMetGroup/loadeR.ECOMS).
}
\details{
The visualization functions are programmed to work with S4 class. This function converts  
 station data to S4 to use those functions.
}
\note{
For gridded data sets with just one ensemble member use as.MrGrid function.
 For gridded data sets with more than one ensemble member use as.MrEnsemble function.
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es}, J. Fernandez and Max Tuni
}
