% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bariatricRat.R
\name{bariatricRat}
\alias{bariatricRat}
\title{NMR spectra of coffee beans from Colombia, Peru and Brazil}
\format{A dataset of 59 rat urine spectra (22913 variables):
\describe{
  \item{metadata}{metadata for each sample}
  \item{X}{NMR spectra}
  \item{ppm}{x axis in ppm}
}}
\source{
{Experimental bariatric surgery in rats generates a cytotoxic chemical environment in the gut contents, Li J, Reshat R, Wu Q, Ashrafian H, Bueter M, Le ROux C, Darzi A et al., Front. Microbiol., 2011}
}
\description{
The data set contains nmr spectra of 59 urine samples of rats.
This dataset is a subset of one published in \url{https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183} and the experimental section can be taken as a reference.
}
\examples{
data("bariatricRat")
}
\references{
\url{https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183}
}
\keyword{bariatric}
\keyword{metabolic,}
\keyword{metabolomics,}
\keyword{rat,}
\keyword{surgery}
\keyword{urine,}
