% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{border_admin0}
\alias{border_admin0}
\title{Haïti border.}
\format{
A sf multiline objet with 1 feature and 6 fields:
\describe{
  \item{fid_1}{fid_1}
  \item{uno}{uno}
  \item{count}{count}
  \item{x_coord}{x_coord}
  \item{y_coord}{y_coord}
  \item{area}{area}
  \item{geometry}{Multiline geometry.}
}
}
\usage{
border_admin0
}
\description{
A multiline shapefile of Haiti's border.
}
\keyword{datasets}
