% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlot.R
\name{bubblePlot}
\alias{bubblePlot}
\title{Bubble plot for visualization of forecast skill of ensemble predictions.}
\usage{
bubblePlot(mm.obj, obs, year.target, detrend = FALSE, score = TRUE,
  size.as.probability = TRUE, piechart = FALSE, only.at = NULL,
  subtitle = NULL, color.reverse = FALSE, pch.neg.score = NULL,
  pch.obs.constant = NULL, pch.data.nan = NULL)
}
\arguments{
\item{mm.obj}{A multi-member list with predictions, either a field or a multi-member 
station object as a result of downscaling of a forecast using station data. See details.}

\item{obs}{The benchmarking observations for forecast verification.}

\item{year.target}{Year within the whole verification period to display the results for. This year is not
included in the computation of the score (operational point of view).}

\item{detrend}{Logical indicating if the data should be detrended. Default is TRUE.}

\item{score}{Logical indicating if the relative operating characteristic skill score (ROCSS) should be included. See 
details. Default is TRUE}

\item{size.as.probability}{Logical indicating if the tercile probabilities (magnitude proportional to bubble radius) 
are drawn in the plot. See details. Default is TRUE.}

\item{piechart}{Logical flag indicating if pie charts should be plot. Default is FALSE.}

\item{only.at}{List with the LonLatCoords of those points selected from the whole grid.}

\item{subtitle}{String to include a subtitle bellow the title. Default is NULL.}

\item{color.reverse}{Logical indicating if the color palete for the terciles (blue, grey, red) should be
reversed (e.g for precipitation). Default is FALSE.}

\item{pch.neg.score}{pch value to highlight the negative score values. Default is NULL. Not available for piecharts.}

\item{pch.obs.constant}{pch value to highlight those whose score cannot be computed due to constant obs 
conditions (e.g. always dry). Default is NULL.}

\item{pch.data.nan}{pch value to highlight those whose score cannot be computed due to time series with all NA values in 
the observations and/or models. Default is NULL. Not available for piecharts.}
}
\description{
Bubble plot for visualization of forecast skill of ensemble predictions. It provides a
 spatially-explicit representation of the skill, resolution and reliability of a probabilistic predictive 
 system in a single map. 
 This function is prepared to plot the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See 
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
For each member, the daily predictions are averaged to obtain a single seasonal forecast. The corresponding terciles 
 for each ensemble member are then computed for the analysis period. Thus, each particular grid point, member and season,
 are categorized into three categories (above, between or below), according to their respective climatological 
 terciles. Then, a probabilistic forecast is computed year by year by considering the number of members falling 
 within each category. For instance, probabilities below 1/3 are very low, indicating that a minority of the members 
 falls in the tercile. Conversely, probabilities above 2/3 indicate a high level of member agreement (more than 66\% of members
 falling in the same tercile). Color represents the tercile with the highest probability for the selected year. The bubble size 
 indicates the probability of that tercile. This option is not plotted if the size.as.probability argument is FALSE.

 Finally, the ROC Skill Score (ROCSS) is computed. For each tercile, it provides a quantitative measure of the forecast skill,
 and it is commonly used to evaluate the performance of probabilistic systems (Joliffe and Stephenson 2003). The value of 
 this score ranges from 1 (perfect forecast system) to -1 (perfectly bad forecast system). A value zero indicates no skill 
 compared with a random prediction. The transparency of the bubble is associated to the ROCSS (negative values are
 plotted with x). This option is not plotted if the score argument is FALSE. The target year is not
 included in the computation of the score (operational point of view).
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es} and J. Fernandez based on the original diagram 
 conceived by Slingsby et al 2009.
}
\references{
Jolliffe, I. T. and Stephenson, D. B. 2003. Forecast Verification: A Practitioner's Guide in Atmospheric 
 Science, Wiley, NY.
 
 Slingsby A., Lowe R., Dykes J., Stephenson D. B., Wood J., Jupp T. E. 2009. A pilot study for the collaborative 
 development of new ways of visualising seasonal climate forecasts. Proc. 17th Annu. Conf. of GIS Research UK, 
 Durham, UK, 1-3 April 2009.
}
\seealso{
Other VisualizeR: \code{\link{as.MrEnsemble}},
  \code{\link{as.MrGrid}}, \code{\link{as.MrStation}},
  \code{\link{convertIntoS4}}, \code{\link{spreadPlot}},
  \code{\link{tercileBarplot}}
}

