% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcRank}
\alias{calcRank}
\title{Calculate a rank}
\usage{
calcRank(id, num_var, cat_var, data, decreasing = FALSE)
}
\arguments{
\item{id}{}

\item{num_var}{}

\item{cat_var}{}

\item{data}{}

\item{decreasing}{}
}
\description{
\code{calcRank}
}

