% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadePlot.R
\name{cascadePlot}
\alias{cascadePlot}
\title{Cascade plot to visualize the climate change modelling uncertainty cascade.}
\usage{
cascadePlot(x, factors, stages = NULL, labels = FALSE,
  label.names = NULL, colors = NULL, title = NULL,
  line.width = NULL, plot.density = TRUE, density.offset = 0,
  density.reverse = TRUE)
}
\arguments{
\item{x}{Numeric vector with the values of the individual projections.}

\item{factors}{List with several factors classifying the data in x.}

\item{stages}{String vector with the names of the factors to be plotted 
from bottom to top.}

\item{labels}{Logical (default = FALSE). Whether to plot individual labels 
for each line. See label.names.}

\item{label.names}{Character vector with the names to be given to each line. 
If NULL (default) names are taken from x.}

\item{colors}{Color vector to be used in the last aggregating step and all 
lines leading to each main branch. If NULL (default) an appropriate rainbow 
scale is used. Its length should be \code{length(levels(factors[[rev(stages)[1]]]))}.}

\item{title}{Title of the plot}

\item{line.width}{Width of the lines.}

\item{plot.density}{Logical (default = TRUE). Add an estimate of the probability
density function (PDF).}

\item{density.offset}{A vector indicating how much the PDF should be shifted 
relative to the bottom part of the plot. Default is 0.}

\item{density.reverse}{Logical (default = TRUE). If TRUE plot the PDF upside down.}
}
\description{
Cascade plot to decompose modelling uncertainty down to individual future climate projections.
}
\examples{
\dontrun{
# This code reproduces Figure 4 (top) from Fernandez et al. (2018)
load(url("http://meteo.unican.es/work/visualizeR/data/multimip_sdelta_es.rda"), verbose=T)
cat(multimip.README)
filter.members <- multimip.factors[["Method"]]=="GCM" & ! multimip.factors[["GCM"]] \%in\% c("CNRM-CM5-r8", "EC-EARTH-r3")
sdeltas <- multimip.prtas.sdeltas["tas",filter.members]
factors <- lapply(multimip.factors, function(x) factor(x[filter.members]))
title <- "Delta T (K)"
cascadePlot(sdeltas, factors, multimip.stages, title=title, density.offset=2)
} 
}
\references{
Hawkins E. 2014. Climate Lab Book. http://www.climate-lab-book.ac.uk/2014/cascade-of-uncertainty

Fernandez, J., Frias, M.D., Cabos, W.D. et al. 2018. Consistency of climate change
projections from multiple global and regional model intercomparison projects. Clim Dyn. https://doi.org/10.1007/s00382-018-4181-8
}
\seealso{
Other visualization functions: \code{\link{bubblePlot}},
  \code{\link{climagram}},
  \code{\link{reliabilityCategories}},
  \code{\link{spreadPlot}}, \code{\link{tercileBarplot}},
  \code{\link{tercilePlot}}
}
\author{
J. Fernandez \email{jesus.fernandez@unican.es} and M.D. Frias based on the
visualization proposed by Ed Hawkins on his Climate Lab Book http://www.climate-lab-book.ac.uk/2014/cascade-of-uncertainty
}
\concept{visualization functions}
