% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroid_admin1}
\alias{centroid_admin1}
\title{Haïti admin 1 centroids shapefile.}
\format{
A sf multipoint object with 10 features and 9 fields:
\describe{
  \item{ADM1_PC}{Admin 1 postal code.}
  \item{ADM1_EN}{Full name in English.}
  \item{ADM1_FR}{Full name in French.}
  \item{ADM1_HT}{Full name in Haitian Creole.}
  \item{ADM0_EN}{Country name in English.}
  \item{ADM0_FR}{Country name in French.}
  \item{ADM0_HT}{Country name in Haitian Creole.}
  \item{ADM0_PC}{Country postal code.}
  \item{ADM1_FR_UPPER}{Admin 1 French name - uppercase.}
  \item{geometry}{Multipoint geometry.}
}
}
\usage{
centroid_admin1
}
\description{
A multipoint shapefile of Haiti's admin 1.
}
\keyword{datasets}
