% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vars_in_df}
\alias{check_vars_in_df}
\title{Check if variables are in data frame}
\usage{
check_vars_in_df(df, vars)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of variable names}
}
\value{
A stop statement
}
\description{
Check if variables are in data frame
}
