% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\name{cleanFun}
\alias{cleanFun}
\title{Data cleanup}
\usage{
cleanFun(id, num_var, cat_var, data)
}
\arguments{
\item{id}{}

\item{num_var}{}

\item{cat_var}{}

\item{data}{}
}
\description{
\code{cleanFun}
}

