% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climagram.R
\name{climagram}
\alias{climagram}
\title{Climagram}
\usage{
climagram(hindcast, forecast = NULL, obs = NULL, year.target = NULL,
  use.anomalies = TRUE, violin = TRUE, add.points = FALSE,
  detrend = FALSE, add.legend = TRUE, clim.time.frame = "none",
  add.eqc.info = TRUE)
}
\arguments{
\item{hindcast}{A climate4R multi-member grid with the hindcast for verification.}

\item{forecast}{A climate4R multi-member grid with the forecast. Default is \code{NULL}. See details.}

\item{obs}{A climate4R grid with the reference observations. They should match the hindcast period.
Default is \code{NULL}.}

\item{year.target}{Year within the hindcast period considered as forecast. Default is NULL. See Details}

\item{use.anomalies}{Should anomalies or the original variable magnitudes be used for data representation? 
Default to \code{TRUE} (i.e., anomalies). See Details.}

\item{violin}{Logical flag indicating whether a violin plot should be added to the graph instead
of an ordinary boxplot. Default is \code{TRUE}.}

\item{add.points}{Logical flag indicating whether crosses indicating the individual forecast members
should be added to the graph. Default to \code{FALSE}.}

\item{detrend}{Logical indicating if the data should be (linearly) detrended. Default is FALSE.}

\item{add.legend}{Should legend be added to the plot. Default to \code{TRUE}}

\item{clim.time.frame}{Default to \code{"none"}. Temporal frame used as reference for the calculation of the climatological
normal. See Details.}

\item{add.eqc.info}{Should EQC (Evaluation and Quality Control) information be added onto the plot?.
Ignored if no \code{obs} is passed. See Details}
}
\value{
Plots a climagram. Note that the number of forecast members falling within each (hindcast) terciles 
are indicated as figures on the right of the forecast box/violins.
}
\description{
A climagram is a box plot for visualization of a seasonal forecast with added EQC information from the hindcast
(and possibly the verifying observations).
}
\details{
\strong{Forecast data}

The product is conceived to display forecast information in an operative framework. However, for
testing purposes, a target year from the hindcast can be extracted via the \code{year.target} argument and 
treated as the forecast year. Note that all the hindcast statistics (including the optional detrending) 
will be computed after extracting the target year from the hindcast.

\strong{Spatial averaging}

 For spatial domains encompassing multiple points/grid points, the spatial average is first computed to obtain a
 unique series for the whole domain vias \code{link[transformeR]{aggregateGrid}} (latitudial weighting is used).

\strong{Forecast/Hindcast representation}
  
 In order to represent the forecast and hindcast information, \dQuote{violins} are used as default,
  although boxplots (see \code{\link[graphics]{boxplot}}) can be obtained by setting 
  argument \code{violin} to \code{FALSE}. For further details on violin construction see
   the documentation of funtion \code{\link[vioplot]{vioplot}}.
 
 Note that the hindcast climatology is calculated by considering the annual multimember averages.

\strong{Using anomalies}

By default, the function will compute anomalies. In this case, the forecast anomalies 
will be calculated considering the hindcast climatology (see next section for details). 

\strong{Climatological reference period}
  
By default, the climatology is computed for the entire seasonal slice of the forecast. For instance,
suppose a forecast for DJF. By default, the DJF anomaly will be computed by subtracting to the 
January forecast the DJF climatology, and so on. If set to \code{"monthly"}, the climatology
is computed sepparately for each month, so the anomalies for January will be computed using the January
 climatology, the February anomalies using the February climatology and so on.
  See argument \code{"time.frame"} of \code{\link[transformeR]{scaleGrid}} for more details.

\strong{EQC Information}  

EQC (Evaluation and Quality Control) information can be optionally added to the plot. This is the default behaviour when 
\code{obs} is supplied. Here CRPS (bias-sensitive) and RPS (bias-insensitive) EQC measures are considered.
}
\note{
All inputs for this function must be at a monthly temporal resolution.

The computation of climatological quantiles requires a representative period to obtain meaningful results.
}
\examples{
\dontrun{
my_load <- function(file.url, verbose = TRUE) {
    tmpfile <- tempfile()
    download.file(url = file.url, destfile = tmpfile)
    readRDS(tmpfile)
}
hindcast <- my_load("http://meteo.unican.es/work/visualizeR/data/tas.cfs.enso34.rds")
forecast <- my_load("http://meteo.unican.es/work/visualizeR/data/tas.cfs.2016.enso34.rds")

# A seasonal forecast product, without reference obs information
climagram(hindcast = hindcast, forecast = forecast, violin = TRUE, obs = NULL)
# Violins can be replaced by ordinary boxplots:
climagram(hindcast = hindcast, forecast = forecast, violin = FALSE, obs = NULL)
# Instead of anomalies, raw values can be used: 
climagram(hindcast = hindcast, forecast = forecast, use.anomalies = FALSE)
# Observations are depicted by the backgroud shadows (median indicated by the dotted line):
# (RPS and CRPS are added by default when observations are included)
obs <- my_load("http://meteo.unican.es/work/visualizeR/data/tas.ncep.enso34.rds")
climagram(hindcast = hindcast, forecast = forecast, obs = obs)
}
}
\seealso{
Other visualization functions: \code{\link{bubblePlot}},
  \code{\link{reliabilityCategories}},
  \code{\link{spreadPlot}}, \code{\link{tercileBarplot}},
  \code{\link{tercilePlot}}
}
\author{
J. Bedia
}
