% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_coffeeSpectra.R
\docType{data}
\name{coffeeSpectra}
\alias{coffeeSpectra}
\title{Spectroscopic information about coffee beans from Colombia, Peru and Brazil}
\format{A data frame with 34 rows and 4 variables:
\describe{
  \item{param}{metadata for each sample}
  \item{irms}{carbon isotope ratio mass spectrometry data (mean of two measurements)}
  \item{nmr}{NMR spectra}
  \item{gc}{intensity of gc spectrogram}
}}
\source{
\url{}
}
\usage{
coffeeSpectra
}
\description{
The data set contains irms, gc and nmr spectra of 34 sample from 3 countries
}
\keyword{datasets}
