\name{colFun}
\alias{colFun}
\title{Add Color to a dataframe}
\usage{
colFun(rang, id, category, data, rank_col = T, top = c(1), flop = c(1),
  nam_blue = "Basler", nam_orange = "TCS", col_neutral = rgb(0, 0, 0,
  alpha = 180, maxColorValue = 255))
}
\arguments{
  \item{rang}{}

  \item{id}{}

  \item{category}{}

  \item{rank_col}{}

  \item{top}{}

  \item{flop}{}

  \item{nam_blue}{}

  \item{nam_orange}{}

  \item{col_neutral}{}
}
\description{
\code{colFun} Add a specific color according to the rank and the
category, i.e., Basler are blue etc.
}

