% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.R
\name{colRank}
\alias{colRank}
\title{Add Color to a dataframe}
\usage{
colRank(rank, id, category, data, var_name = "color", rank_col = TRUE,
  top = c(1), flop = c(1), nam_1 = NULL, nam_2 = NULL, col_1 = NULL,
  col_2 = NULL, col_red = rgb(139, 0, 0, alpha = 255, maxColorValue = 255),
  col_green = rgb(0, 100, 0, alpha = 255, maxColorValue = 255),
  col_neutral = rgb(0, 0, 0, alpha = 255, maxColorValue = 255),
  col_NA = "black")
}
\arguments{
\item{rank}{variable with the indiviudual rank values}

\item{id}{ID variable for the individual data points}

\item{category}{variable that defines the bins}

\item{rank_col}{True or False variable to set if the rank colors are shown}

\item{top}{Variable to define how many profiles are in the top section}

\item{flop}{Variable to define how many profiles are in the flop section}

\item{nam_1}{Variable to color an individual category}

\item{nam_2}{Variable to color an individual category}

\item{col_1}{color for category 1}

\item{col_2}{color for category 2}

\item{col_red}{red color}

\item{col_green}{green color}

\item{col_neutral}{neutral color}

\item{col_NA}{color for the missing variables}
}
\description{
\code{colFun} Add a specific color according to the rank and the
category, i.e., Basler are blue etc.
}

