% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_impact.R
\name{cols_impact}
\alias{cols_impact}
\title{Function to extract IMPACT colors as hex codes}
\usage{
cols_impact(..., unnamed = TRUE)
}
\arguments{
\item{...}{Character names of reach colors. If NULL returns all colors}

\item{unnamed}{Should the output vector be unnamed? Default to `TRUE`}
}
\value{
An hex code or hex codes named or unnamed
}
\description{
Function to extract IMPACT colors as hex codes
}
\details{
This function needs to be modified to add colors
}
