% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFun.R
\name{convertIntoS4}
\alias{convertIntoS4}
\title{Convert a dataset to a S4 class.}
\usage{
convertIntoS4(obj)
}
\arguments{
\item{obj}{A list the with the elements Variable, Dates, Data, xyCoords and Members 
and InitializationDates (for gridded data) or MetaData (for station data)}
}
\description{
Convert a data set to a S4 class. The data set must have the elements Variable, 
 Dates, Data, xyCoords. Depending of the kind of data it must also have the elements Members 
 and InitializationDates (for gridded data) or MetaData (for station data). This function is 
 prepared to convert the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See 
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
Most of the visualization functions are programmed to work with S4 class. This function converts  
 datasets into S4 to use those functions.
}
\note{
For gridded data sets with just one ensemble member use as.MrGrid function.
 For station data use as.MrStation function.
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es} and J. Fernandez
}
