% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createView.R
\name{createView}
\alias{createView}
\title{This function creates a blank view list object}
\usage{
createView(name = "visualizeR")
}
\arguments{
\item{name}{the name of the view that will appears in the browser}
}
\value{
a blank view list object
}
\description{
This function creates a blank view list object
}
\examples{
createView()
}
