% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cutTxt}
\alias{cutTxt}
\title{Cut down text to required id}
\usage{
cutTxt(input, identifier, cut = F, regex = "[[:alnum:]]{1, }")
}
\arguments{
\item{input}{Data}

\item{identifier}{}

\item{cut}{}

\item{regex}{}
}
\description{
\code{cutTxt} Cuts down the text to a required id
}

