% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_selection.R
\name{dataframe_selection}
\alias{dataframe_selection}
\title{Dataframe Selection Module}
\usage{
dataframe_selection(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny parameters.}
}
\value{
A reactive value holding the name of the selected dataframe.
}
\description{
This function handles the display of a modal dialog to select a dataframe
from the Global Environment and returns the selected dataframe.
}
\keyword{internal}
