% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density plots for climatological series}
\usage{
densityPlot(..., h.lines = NULL, v.lines = NULL, group.index = NULL,
  lonLim = NULL, latLim = NULL, dplot.custom = list())
}
\arguments{
\item{...}{Input grids (or station data). It can be a list of grids too.}

\item{h.lines}{Numeric sequence indicating the position of dashed horizontal lines.}

\item{v.lines}{Numeric sequence indicating the position of dashed vertical lines. The default (NULL),
plots vertical lines in the mean value of the probability density function.}

\item{group.index}{Character or numeric passed to argument group in density plot.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, 
in decimal degrees, of the bounding box selected. If NULL (default), the whole 
longitudinal range is selected}

\item{latLim}{Same as lonLim, but for the selection of the latitudinal range.}

\item{dplot.custom}{List of arguments as passed to \code{\link[lattice]{densityplot}}. Arguments \code{panel} cannot 
be modified, thus, if specified, it will be ignored. 
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{dplot}.}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
A wrapper for the lattice density plot methods for grid and station data.
}
\author{
M. Iturbide
}
