% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFun.R
\name{detrend.data}
\alias{detrend.data}
\title{Detrend data from a S4 object}
\usage{
detrend.data(obj, obj2 = NULL)
}
\arguments{
\item{obj}{S4 object with dimensions c("var", "member", "time", "y", "x") it can be a station, field, multi-member field, etc}

\item{obj2}{obj2. Default to \code{NULL}.}
}
\value{
A S4 object with the data detrended. The object has dimensions c("var", "member", "time", "y", "x")
}
\description{
Detrend data from a S4 object with the dimensions c("var", "member", "time", "y", "x")
}
\author{
M. D. Frias \email{mariadolores.frias@unican.es}, J. Fernandez and J. Bedia
}

