% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposeData.R
\name{exposeData}
\alias{exposeData}
\title{This function parse the datasets present in the data list object and configure the view list object accordingly.
For the view to be able to access the data in the https://github.com/npellet/visualizer it is necessary fot the data to
be exposed.}
\usage{
exposeData(view, data)
}
\arguments{
\item{view}{view list object that should be modified}

\item{data}{data list object with datasets to be exposed}
}
\value{
a modified view list object with correct variable configuration
}
\description{
This function parse the datasets present in the data list object and configure the view list object accordingly.
For the view to be able to access the data in the https://github.com/npellet/visualizer it is necessary fot the data to
be exposed.
}
\examples{
#reset variables in view
v$variables <- list()

#create new variables
v <- exposeData(view=v, data=d)
}
