% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frontier_admin0}
\alias{frontier_admin0}
\title{Haïti frontier with Dominican Republic.}
\format{
A sf multipoint objet with 4 features and 8 fields:
\describe{
  \item{fid_1}{fid_1}
  \item{objectid}{objectid}
  \item{id}{id}
  \item{fromnode}{fromnode}
  \item{tonode}{tonode}
  \item{leftpolygo}{leftpolygo}
  \item{rightpolygo}{rightpolygo}
  \item{shape_leng}{shape_leng}
  \item{geometry}{Multiline geometry.}
}
}
\usage{
frontier_admin0
}
\description{
A multiline shapefile of Haiti's frontier with Dominican Republic.
}
\keyword{datasets}
