% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{\%notallin\%}
\alias{\%notallin\%}
\title{Not All In Operator}
\usage{
a \%notallin\% b
}
\arguments{
\item{a}{Vector to test}

\item{b}{Vector to test against}
}
\value{
TRUE if at least one element of `a` is not in `b`, otherwise FALSE
}
\description{
Tests if not all elements of `a` are contained in `b`.
}
