% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In Operator}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{Vector or value to test}

\item{b}{Vector to test against}
}
\value{
Logical vector with TRUE for elements of `a` that are not in `b`
}
\description{
A negation of the `%in%` operator that tests if elements of `a` are not in `b`.
}
