% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbar.R
\name{hbar}
\alias{hbar}
\title{Simple horizontal bar chart}
\usage{
hbar(
  .tbl,
  x,
  y,
  group = NULL,
  initiative = "reach",
  x_title = "",
  y_title = "",
  group_title = NULL,
  font_family = "Leelawadee",
  stack = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{Some data}

\item{x}{Some numeric column on the x scale}

\item{y}{Some column on the y scale}

\item{group}{Some grouping categorical column, e.g. administrative areas}

\item{initiative}{Either "reach" or "agora" or "impact" for the color palette}

\item{x_title}{The x scale title. Default to empty string}

\item{y_title}{The y scale title. Default to empty string}

\item{group_title}{The group legend title. Defaut to NULL}

\item{font_family}{The font family. Default to "Leelawadee"}

\item{stack}{Should the chart be stacked? Default to "FALSE" (dodge)}

\item{...}{Other arguments to be passed to "simplevis::gg_hbar" or "simplevis:gg_hbar_col"}
}
\value{
A horizontal bar chart
}
\description{
without any change to the x scale
}
