% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{hindcast2prob}
\alias{hindcast2prob}
\title{internal function prob2bin}
\usage{
hindcast2prob(hindcast, n.events, hindcast4cats = NULL)
}
\arguments{
\item{hindcast}{3D-array of predictions, dimensions = (member, time, npoints)}

\item{n.events}{Number of categories (3 for terciles)}

\item{hindcast4cats}{Optional. 3D-array of predictions for which calculate the categories (e.g., terciles)}

\item{dimensions}{(member, time, npoints)}
}
\value{
hindcastprob: 3D-array of probabilistic predictions, dimensions = (n.events, time, npoints)
}
\description{
This function provides the object needed by "calculateReliability_v2.R" 
for calculating the reliability categories of a probabilistic prediction.
}
\note{
For hindcast4cats, categories are calculated at model- (not at member-) level
}
\author{
R. Manzanas \& M.Iturbide
}
\keyword{internal}
