% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_admin1}
\alias{indicator_admin1}
\title{Indicator admin 1 polygons shapefile.}
\format{
A sf multipoint object with 10 features and 10 fields:
\describe{
  \item{ADM1_PC}{Admin 1 postal code.}
  \item{admin1}{Admin 1 unique id.}
  \item{opn_dfc}{Proportion of HHs that reported open defecation as sanitation facility.}
  \item{ADM1_EN}{Full name in English.}
  \item{ADM1_FR}{Full name in French.}
  \item{ADM1_HT}{Full name in Haitian Creole.}
  \item{ADM0_EN}{Country name in English.}
  \item{ADM0_FR}{Country name in French.}
  \item{ADM0_HT}{Country name in Haitian Creole.}
  \item{ADM0_PC}{Country postal code.}
  \item{geometry}{Multipolygon geometry.}
}
}
\usage{
indicator_admin1
}
\description{
A multipolygon shapefile of Haiti's admin 1 with an indicator column 'opn_dfc'.
}
\keyword{datasets}
