% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize session-specific variables}
\usage{
init(session)
}
\arguments{
\item{session}{The Shiny session object.}
}
\value{
None.
}
\description{
This function initializes variables to be used in the Shiny session.
}
\keyword{internal}
