% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_get.R
\name{input_get}
\alias{input_get}
\title{input_get}
\usage{
input_get(input, id, parameter)
}
\arguments{
\item{input}{Standard Shiny parameters}

\item{id}{ID of geom}

\item{parameter}{Name of the parameter}
}
\value{
Value of the parameter
}
\description{
Get value of an input
}
