% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_list.R
\name{input_list}
\alias{input_list}
\title{input_list}
\usage{
input_list(id, param, label = "", choices, mode = "both")
}
\arguments{
\item{id}{Panel ID}

\item{param}{Name of the parameter which this input control}

\item{label}{Label above the input}

\item{choices}{Vector of choices in the list}

\item{mode}{Mode where parameter is used
\itemize{
\item \code{"both"}: By default, widget appear in both mode
\item \code{"advanced"}: Only advanced mode
\item \code{"basic"}: Only basic mode
}}
}
\value{
UI List
}
\description{
Basic list input
}
