% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_numeric.R
\name{input_numeric}
\alias{input_numeric}
\title{input_numeric}
\usage{
input_numeric(id, param, label = "", default = 0, mode = "both")
}
\arguments{
\item{id}{Panel ID}

\item{param}{Name of the parameter which this input control}

\item{label}{Label above the input}

\item{default}{Default value of the input}

\item{mode}{Mode where parameter is used
\itemize{
\item \code{"both"}: By default, widget appear in both mode
\item \code{"advanced"}: Only advanced mode
\item \code{"basic"}: Only basic mode
}}
}
\value{
UI Numeric selector
}
\description{
Basic numeric input
}
