% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{LaunchApp}
\alias{LaunchApp}
\title{Launch the Shiny Application}
\usage{
LaunchApp()
}
\value{
A running Shiny app instance.
}
\description{
This function launches the Shiny application by combining the user interface (UI)
and the server logic into a single app. It serves as the entry point for the app.
}
\examples{
if (interactive()) {
  LaunchApp()
}
}
