% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{layer_base}
\alias{layer_base}
\title{Layer base}
\usage{
layer_base(type, id = NULL, drag = T, session = getDefaultReactiveDomain())
}
\arguments{
\item{type}{String contain type of the layer}

\item{id}{string ID of the layer}

\item{drag}{Box is dragable ?}

\item{session}{Standard Shiny parameters}
}
\value{
UI button
}
\description{
Constructor for the layers
}
\keyword{internal}
