% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_select_new.R
\name{layer_select_new}
\alias{layer_select_new}
\title{layer_select_new}
\usage{
layer_select_new()
}
\value{
A modalDialog object with custom content.
}
\description{
Modal window UI where user can select the type of new layer
}
\keyword{internal}
