% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerBar.R
\name{layerBar}
\alias{layerBar}
\title{UI for the left sidebar used to select layer}
\usage{
layerBar()
}
\value{
Shiny UI Object
}
\description{
UI for the left sidebar used to select layer
}
\keyword{internal}
