% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerBarEvents.R
\name{layerBarEvents}
\alias{layerBarEvents}
\title{Layer Bar Event Handlers}
\usage{
layerBarEvents(input, output, session, id_layer_list)
}
\arguments{
\item{input, }{output, session Standard Shiny parameters.}

\item{id_layer_list}{}
}
\description{
Encapsulates the logic to handle the layer bar interaction.
}
\keyword{internal}
