% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerBox.R
\name{layerBox}
\alias{layerBox}
\title{Layer Button}
\usage{
layerBox(type, id = NULL, drag = T, session = getDefaultReactiveDomain())
}
\arguments{
\item{type}{Type of the layer}

\item{drag}{Dragable box}
}
\value{
UI button
}
\description{
Layer Button
}
\keyword{internal}
