% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{line_admin1}
\alias{line_admin1}
\title{Haïti admin 1 lines shapefile.}
\format{
A sf multiline object with 10 features and 8 fields:
\describe{
  \item{ADM1_EN}{Full name in English.}
  \item{ADM1_FR}{Full name in French.}
  \item{ADM1_HT}{Full name in Haitian Creole.}
  \item{ADM0_EN}{Country name in English.}
  \item{ADM0_FR}{Country name in French.}
  \item{ADM0_HT}{Country name in Haitian Creole.}
  \item{ADM0_PCODE}{Country postal code.}
  \item{geometry}{Multiline geometry.}
}
}
\usage{
line_admin1
}
\description{
A multiline shapefile of Haiti's admin 1.
}
\keyword{datasets}
