% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliaryFun.R
\name{MrQuantile}
\alias{MrQuantile}
\title{Get quantiles from a S4 object}
\usage{
MrQuantile(obj, k = NULL)
}
\arguments{
\item{obj}{S4 object with dimensions c("var", "member", "time", "y", "x") it can be a station, field, multi-member field, etc}

\item{k}{order of the quantile/s (e.g. c(1/3, 2/3) for terciles)}
}
\value{
A S4 object with quantiles. The object has dimensions c("var", "member", "time", "y", "x")
}
\description{
Get quantiles from a S4 object with the dimensions c("var", "member", "time", "y", "x")
}
\author{
M. D. Frias \email{mariadolores.frias@unican.es} and J. Fernandez
}

