% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{nboxplot.stats}
\alias{nboxplot.stats}
\title{boxplot.stats with set probs}
\usage{
nboxplot.stats(
  x,
  coef = 1.5,
  do.conf = TRUE,
  do.out = FALSE,
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9)
)
}
\arguments{
\item{x}{numeric vector}

\item{coef}{see boxplot.stats}

\item{do.conf}{see boxplot.stats}

\item{do.out}{see boxplot.stats}

\item{probs}{numeric with probabilites between 0 and 1}
}
\description{
boxplot.stats with set probs
}
\author{
M. Iturbide
}
