% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{obs2bin}
\alias{obs2bin}
\title{internal function obs2bin}
\usage{
obs2bin(obs, n.events)
}
\arguments{
\item{obs}{2D-matrix of observations, dimensions = (time, npoints)}

\item{n.events}{number of categories (3 for terciles)}
}
\value{
bincat: list with two elements:  
bin: 3D-array of binary (0/1) observations, dimensions = (n.events, time, npoints)
cat: 2D-matrix with the observed category, dimensions = (time, npoints)
}
\description{
This function provides the object needed by "calculateReliability_v2.R" 
for calculating the reliability categories of a probabilistic prediction.
}
\author{
R. Manzanas \& M.Iturbide
}
\keyword{internal}
