% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{out}
\alias{out}
\title{Clean up session-specific variables}
\usage{
out(session)
}
\arguments{
\item{session}{The Shiny session object.}
}
\value{
None.
}
\description{
This function cleans up when the Shiny session ends.
}
\keyword{internal}
