% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_fallback.R
\name{pal_fallback}
\alias{pal_fallback}
\title{Return function to interpolate a fallback palette base on viridis::magma()}
\usage{
pal_fallback(
  reverse = FALSE,
  color_ramp_palette = FALSE,
  discrete = FALSE,
  n = 5,
  ...
)
}
\arguments{
\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{color_ramp_palette}{Should the output be a `grDevices::colorRampPalette` function or a vector of hex codes? Default to the latter with `FALSE`}

\item{discrete}{Boolean. Discrete or not? Default to FALSE.}

\item{n}{Number of colors in the palette. Default to 5. Passe to `viridis::magma()`}

\item{...}{Other parameters to pass to `grDevices::colorRampPalette()`}
}
\value{
A color palette
}
\description{
Return function to interpolate a fallback palette base on viridis::magma()
}
