% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette}
\alias{palette}
\title{Interpolate a color palette}
\usage{
palette(palette = "cat_5_main", reverse = FALSE, show_palettes = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette in palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{show_palettes}{Should the ouput be the set of palettes names to pick from? Default to `FALSE`}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A color palette
}
\description{
Interpolate a color palette
}
