% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gen.R
\name{palette_gen}
\alias{palette_gen}
\alias{palette_gen_categorical}
\alias{palette_gen_sequential}
\title{Generate color palettes}
\usage{
palette_gen(palette, type, direction = 1, ...)

palette_gen_categorical(palette = "branding_reach", direction = 1)

palette_gen_sequential(palette = "seq_7_artichoke", direction = 1, ...)
}
\arguments{
\item{palette}{Palette name from [palette()].}

\item{type}{"categorical" or "sequential" or "divergent".}

\item{direction}{1 or -1; should the order of colors be reversed?}

\item{...}{Additional arguments to pass to [colorRampPalette()] when type is "continuous".}
}
\description{
[palette_gen()] generates a color palette and let you choose whether continuous or discrete. [palette_gen_categorical()] and [palette_gen_sequential()] generates respectively discrete and continuous palettes.
}
