% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_create.R
\name{panel_create}
\alias{panel_create}
\title{panel_create}
\usage{
panel_create(type, id_panel)
}
\arguments{
\item{type}{String of layer's type}

\item{id_panel}{String of layer's ID}
}
\value{
Div of a parameters panel
}
\description{
Constructor of the parameters panel
}
