% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_theme_opt.R
\name{panel_theme_opt}
\alias{panel_theme_opt}
\title{panel_line}
\usage{
panel_theme_opt(id)
}
\arguments{
\item{id}{String of layer's ID}
}
\value{
Div of a global theme option panel
}
\description{
Constructor of global theme option panel
}
