% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramBar.R
\name{paramBar}
\alias{paramBar}
\title{UI for the left sidebar used to control parameters}
\usage{
paramBar()
}
\value{
Shiny UI Object
}
\description{
UI for the left sidebar used to control parameters
}
\keyword{internal}
