% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display.R
\name{plot_display}
\alias{plot_display}
\title{GGplot renderer}
\usage{
plot_display(input)
}
\arguments{
\item{input}{Standard Shiny parameters}
}
\value{
GGplot2 plot
}
\description{
Dynamic renderer of the plot
}
\keyword{internal}
