% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{plot_line}
\usage{
plot_line(param, id, plot)
}
\arguments{
\item{param}{List of parameters value for this geom}

\item{id}{ID of the geom}

\item{plot}{GGplot on which to add the line.}
}
\value{
GGplot with a new line geom
}
\description{
Create a line on the plot (h, v or ab)
}
