% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDisplay.R
\name{plotDisplay}
\alias{plotDisplay}
\title{GGplot renderer}
\usage{
plotDisplay(input)
}
\arguments{
\item{input}{Standard Shiny parameters}
}
\value{
GGplot2 plot
}
\description{
GGplot renderer
}
\keyword{internal}
