% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotMap}
\alias{plotMap}
\title{Plot data on a shape file (Geovisualisation)}
\usage{
plotMap(shape, path_png, var, labels = T, perc = F, col_pos = "Greens",
  col_neg = "Reds", col_zero = "#999999", col_missing = "black",
  numInterval = 10, numColLength = 4)
}
\arguments{
\item{shape}{Data that has to be merged with the shape file.}

\item{path_png}{Output path where the .png file is saved.}

\item{var}{Name of the variable to be plotted.}

\item{labels}{Logical value to determine if labels are put onto the plot.}

\item{param}{Logical value to determine if the label values are transformed into percentages.}

\item{col_shades}{What kind of colors shades are used.}

\item{col_zero}{The color type for zeros.}

\item{col_missing}{The color type for missings (NA's).}
}
\description{
\code{plotMap} plots datapoints on a shape file.
}
\examples{
plotMap(shp_obj, '/plots/schaeden.png', 'Anzahl')

plotMap(shp_obj, '/plots/kunden.png', 'Anzahl', perc = T, col_shades = 'Blues')
}

