% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Simple bar chart}
\usage{
point(
  df,
  x,
  y,
  group = NULL,
  flip = TRUE,
  alpha = 1,
  size = 1,
  x_title = NULL,
  y_title = NULL,
  group_title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  theme = theme_reach()
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric column.}

\item{y}{A character column or coercible as a character column.}

\item{group}{Some grouping categorical column, e.g. administrative areas or population groups.}

\item{flip}{TRUE or FALSE. Default to TRUE or horizontal bar plot.}

\item{alpha}{Fill transparency.}

\item{size}{Point size.}

\item{x_title}{The x scale title. Default to NULL.}

\item{y_title}{The y scale title. Default to NULL.}

\item{group_title}{The group legend title. Default to NULL.}

\item{title}{Plot title. Default to NULL.}

\item{subtitle}{Plot subtitle. Default to NULL.}

\item{caption}{Plot caption. Default to NULL.}

\item{theme}{Whatever theme. Default to theme_reach().}
}
\value{
A bar chart
}
\description{
Simple bar chart
}
