% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point_reach}
\alias{point_reach}
\title{Simple point chart}
\usage{
point_reach(
  df,
  x,
  y,
  group = NULL,
  palette = "main",
  reverse = FALSE,
  family = "Leelawadee",
  alpha = 1,
  size = 1.5,
  x_title = NULL,
  y_title = NULL,
  group_title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  text_size = 10,
  title_size = 14,
  legend_position = "right",
  ...
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric column.}

\item{y}{A character column or coercible as a character column.}

\item{group}{Some grouping categorical column, e.g. administrative areas or population groups.}

\item{palette}{Palette name from 'pal_reach()'.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{family}{The font family for all plot's texts. Default to "Leelawadee".}

\item{alpha}{Transparency.}

\item{size}{Dot size. Default to 1.5.}

\item{x_title}{The x scale title. Default to NULL.}

\item{y_title}{The y scale title. Default to NULL.}

\item{group_title}{The group legend title. Default to NULL.}

\item{title}{Plot title. Default to NULL.}

\item{subtitle}{Plot subtitle. Default to NULL.}

\item{caption}{Caption title string. Default to NULL.}

\item{text_size}{The size of all text other than the title, subtitle and caption. Defaults to 10.}

\item{title_size}{The size of the title text. Defaults to 14.}

\item{legend_position}{Position of the legend; Default to "right". Can take "right", "left", "top", "bottom" or "none".}

\item{...}{Other arguments to be passed to "ggblanket::gg_col"}
}
\value{
A bar chart
}
\description{
`ggblanket` as internals for deciding whether the bar chart is horizontally readable.
}
