% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{PRUDENCEregions}
\alias{PRUDENCEregions}
\title{PRUDENCE regions.}
\format{A SpatialPolygons object}
\description{
SpatialPolygons object of the geographical regions used in the PRUDENCE project.

For further detail, see \url{http://ensemblesrt3.dmi.dk/quicklook/regions.html}
}
\examples{
\dontrun{
data(PRUDENCEregions)
plot(PRUDENCEregions)
}
}

