% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Pushes new view.json or data.json to the visualization server}
\usage{
push(visualization, type, list)
}
\arguments{
\item{visualization}{the visualization to be updated}

\item{type}{the type of file to be update, view or data}

\item{list}{the list object that contains the view or data}
}
\value{
an updated visualization object and new data/view.json on the server
}
\description{
Pushes new view.json or data.json to the visualization server
}
\examples{
example
saveJSON(v, "~/view.json")
saveJSON( object=exposeData(view=v, data=d), pathToFile="~/view.json")
}
