% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{reliabilityCategories}
\alias{reliabilityCategories}
\title{Reliability categories of a probabilistic prediction.}
\usage{
reliabilityCategories(obs, prd, nbins = 3, nbinsprob = 100, nboot = 100,
  sigboot = 0.05, diagrams = TRUE)
}
\arguments{
\item{obs}{Grid of observations}

\item{prd}{Grid of predictions}

\item{nbins}{(optional): number of categories considered (e.g. 3 for terciles). By default nbins = 3}

\item{nbinsprob}{(optional): number of probability bins considered. By default nbinsprob = 10}

\item{nboot}{number of samples considered for bootstrapping. By default nboot = 100}

\item{sigboot}{Optional. Confidence interval for the reliability line. By default sigboot = 0.05}

\item{diagrams}{Logical (default = TRUE). Plotting results.}
}
\value{
Same as prd but with an aditional list ($ReliabilityCategories) containing the following elements:
catcol: color of the reliability category
catname: reliability category
sl: slope of the reliability line
sl_lower: lower bound confidence for slope (according to "sigboot")
sl_lower: upper bound confidence for slope (according to "sigboot")
}
\description{
Calculates (and draws)
reliability diagrams and the related reliability categories, according to Weisheimer et al. 2014.
}
\author{
R. Manzanas \& M.Iturbide
}
\references{
Weisheimer, A., Palmer, T.N., 2014. On the reliability of seasonal climate forecasts. Journal of The Royal Society Interface 11, 20131162. doi:10.1098/rsif.2013.1162
}

