% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.col.R
\name{rep.col}
\alias{rep.col}
\title{repeat a column when needed to match other matrix dimension (not exported)}
\usage{
\method{rep}{col}(x, n)
}
\arguments{
\item{x}{a vector to be repeated}

\item{n}{number ot time to repeat the vector}
}
\value{
a matrix
}
\description{
repeat a column when needed to match other matrix dimension (not exported)
}
\examples{
rep.col(c(1,2,3), 2)
}
