% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.row.R
\name{rep.row}
\alias{rep.row}
\title{repeat a row when needed to match other matrix dimension (not exported)}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{a vector to be repeated}

\item{n}{number ot time to repeat the vector}
}
\value{
a matrix
}
\description{
repeat a row when needed to match other matrix dimension (not exported)
}
\examples{
rep.row(c(1,2,3), 2)
}
