% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveJSON.R
\name{saveJSON}
\alias{saveJSON}
\title{Converts the list objects into json files}
\usage{
saveJSON(object, pathToFile)
}
\arguments{
\item{object}{the view list object to be saved}

\item{pathToFile}{the path to file}
}
\value{
a json file to be passed to the https://github.com/npellet/visualizer
}
\description{
Converts the list objects into json files
}
\examples{
example
saveJSON(v, "~/view.json")
saveJSON( object=exposeData(view=v, data=d), pathToFile="~/view.json")
}
