% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_visualizer_discrete}
\alias{scale_visualizer_discrete}
\title{One scale for all}
\usage{
scale_visualizer_discrete(
  palette = "cat_5_main",
  direction = 1,
  reverse_guide = TRUE,
  ...
)
}
\arguments{
\item{palette}{Palette name from [palette()].}

\item{direction}{1 or -1; should the order of colors be reversed?}

\item{reverse_guide}{Boolean indicating whether the guide should be reversed.}

\item{...}{Additional arguments passed to [ggplot2::discrete_scale()] if discrete or [ggplot2::scale_fill_gradient()] if continuous.}
}
\description{
This function is based on [palette()]. If palette is NULL, the used palette will be magma from gpplot2's viridis scale constructors.
}
