% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side.density.R
\name{side.density}
\alias{side.density}
\title{Add an estimate of the probability density function.}
\usage{
side.density(x, horiz = TRUE, col = "black", scale = 0.1, alpha = 0.3,
  shades = FALSE, prob = c(0.25, 0.5, 0.75), density.offset = 0,
  density.reverse = FALSE)
}
\arguments{
\item{x}{Numeric vector with the values of the individual projections.}

\item{horiz}{Logical (default = TRUE). If TRUE the PDF is drawn horizontally.}

\item{col}{Color for the PDF.}

\item{scale}{A vector indicating how much to scale the PDF. Default is 0.1}

\item{alpha}{A vector indicating the transparency to be applied to col. 
Alpha ranges from 0 (fully transparent) to 1 (opaque). Default is 0.3.}

\item{shades}{Logical (default = FALSE). If TRUE add shades for the quantiles 
defined in prob.}

\item{prob}{A vector giving the quantiles to be highlighted in the PDF. 
Default is the quartiles.}

\item{density.offset}{A vector indicating how much the PDF should be shifted 
relative to the bottom part of the plot. Default is 0.}

\item{density.reverse}{Logical (default = FALSE). If TRUE plot the PDF upside down.}
}
\description{
The function adds an estimate of the probability density function (PDF) 
  of the projections.
}
\details{
The function is intended for internal use only. The function adds the PDF to the cascade plot.
}
\author{
J. Fernández and M.D. Frías
}
\keyword{internal}
