% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skillMap.R
\name{skillMap}
\alias{skillMap}
\title{A wrapper of \code{spatialPlot} for the creation of verification maps for seasonal forecast systems}
\usage{
skillMap(easyVeriGrid, stippling = list(threshold = NULL, condition = NULL),
  stippling.point.options = NULL, backdrop.theme = NULL, title = NULL,
  ...)
}
\arguments{
\item{easyVeriGrid}{A climatological grid with a verification output. See details}

\item{stippling}{A key-value list of arguments passed to \code{map.stippling}: 
\code{threshold} and \code{condition}. Ignored by default, returning a map without stippling.}

\item{stippling.point.options}{Default to \code{NULL}. Further graphical arguments passed to points 
(e.g. \code{cex}, \code{pch} etc.)}

\item{backdrop.theme}{See \code{\link{spatialPlot}}}

\item{title}{Title of the plot}

\item{...}{Further graphical options passed to \code{\link{spatialPlot}}.}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
A wrapper of \code{spatialPlot} for the creation of verification maps for seasonal forecast systems.
 It provides a convenient interface for \code{\link{map.stippling}}
}
\details{
The function applies the \code{\link{spatialPlot}} function, that in turn uses \code{lattice-methods}.

\strong{Graphical options}

Some examples of specific map graphical options are available in the help of function \code{\link[sp]{spplot}}.
 In addtion, fine-tuning of the resulting plots can be obtained using the arguments of \pkg{lattice} plots. For
 an overview, see the help of function \code{\link[lattice]{xyplot}}.
}
\seealso{
The bridging function \code{\link[transformeR]{easyVeri2grid}} from package \pkg{transformeR} allows for 
the conversion of verification outputs from package \pkg{easyVerification} to the \code{climate4R} data structure
used by the function.
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{\link[sp]{spplot}}.
}
\author{
J. Bedia
}
