% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripePlot.R
\name{stripePlot}
\alias{stripePlot}
\title{Lattice plot methods for producing stripe plots}
\usage{
stripePlot(
  grid,
  aggr.spatial = list(FUN = "mean", na.rm = TRUE),
  color.theme = "RdBu",
  rev.colors = FALSE,
  xlabels = NULL,
  ylabels = NULL,
  ...
)
}
\arguments{
\item{grid}{Input grid (or station data), ideally multimember and with time dimension > 1.}

\item{aggr.spatial}{list containing the function and the related arguments to perform spatial 
aggregation. Default is \code{list(FUN = mean, na.rm = TRUE)}.}

\item{color.theme}{A character string indicating the color theme to use in the map. 
Valid values are those available in the \code{\link{RColorBrewer}} themes. Additionally,
the \code{"jet.colors"} palette can be used (the rainbow colors, in general not advised, though),
for backwards compatibility. Default to the diverging, colorblind-friendly \code{"RdYlBu"} palette.
NOTE: the \code{color.theme} argument will be overriden if the \code{col.regions} option from \code{levelplot} is used.}

\item{rev.colors}{Should the chosen color theme be reversed? (default to \code{FALSE},
leaving the palette \dQuote{as is}).}

\item{xlabels}{Character string of the labels to be shown in the x axis (Default to NULL)}

\item{ylabels}{Character string of the labels to be shown in the y axis (Default to NULL)}

\item{...}{Additional arguments as passed to function \code{levelplot}.}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
A wrapper for the lattice (trellis) plot methods for grid and station data.
}
\details{
The function applies the \code{\link[lattice]{levelplot}} method after computing spatial aggregation 
(parameter \code{aggr.spatial}) to the imput grids (or station data).
 
This function is appropriate for multimember grids, each row represents a member.
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{levelplot}. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link[lattice]{trellis.par.set}}.
}
\examples{
\donttest{
require(climate4R.datasets)
data("CFS_Iberia_pr")
stripePlot(CFS_Iberia_pr)
}
}
\author{
M. Iturbide
}
