% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\name{tas.cfs.operative.2016}
\alias{tas.cfs.operative.2016}
\title{CFSv2 operative seasonal forecast of global DJF mean temperature (2016)}
\format{A grid}
\source{
Dowloaded from the ECOMS User Data Gateway \url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
CFSv2 operative seasonal forecast of DJF mean temperature for the entire globe of year 2016
 (i.e. December 2015, Januery-February 2016). Data are annually aggregated. Ensemble of 24 members,
  corresponding to the lead month 1 predictions (November 2015 initializations).
}
\examples{
\dontrun{
data(tas.cfs.operative.2016)
require(transformeR)
plotClimatology(climatology(tas.cfs.operative.2016, parallel = TRUE), backdrop.theme = "coastline")
}
}
\references{
Saha, S. et al. 2014. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}

