% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{tas.ncep}
\alias{tas.ncep}
\title{NCEP/NCAR reanalysis DJF mean temperature for the entire globe}
\format{A grid}
\source{
Dowloaded from the ECOMS User Data Gateway \url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
NCEP/NCAR reanalysis of December to February (DJF) mean temperature for the entire globe (1983-2010). Data are annually aggregated.
}
\examples{
data(tas.ncep)
require(transformeR)
plotClimatology(climatology(tas.ncep), backdrop.theme = "coastline")
}
