% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileMap.R
\name{tercileMap}
\alias{tercileMap}
\title{Tercile map for visualization of forecast skill of seasonal climate predictions.}
\usage{
tercileMap(hindcast, forecast, ...)
}
\arguments{
\item{hindcast}{A multi-member list with the hindcast for verification. See details.}

\item{forecast}{A multi-member list with the forecasts. Default is NULL.}

\item{...}{Optional arguments passed to \code{\link{spatialPlot}}}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
Tercile map for the visualization of forecast skill of seasonal climate predictions.
 This function is prepared to plot the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See 
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
The function applies the \code{\link{spatialPlot}} function, that in turn uses \code{lattice-methods}.
}
\note{
The function is based on the tercile maps delivered by
C3S seasonal forecasts (\url{http://climate.copernicus.eu/s/charts/c3s_seasonal/}).
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{\link[sp]{spplot}}.
}
\examples{
library(transformeR)
hindcast <- subsetGrid(CFS_Iberia_tas, years = 1983:2001)
forecast <- subsetGrid(CFS_Iberia_tas, years = 2002)
tercileMap(hindcast, forecast)
}
\author{
M. Iturbide
}
