% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tercilePlot.R
\name{tercilePlot}
\alias{tercilePlot}
\title{Tercile plot for visualization of forecast skill of ensemble predictions.}
\usage{
tercilePlot(hindcast, obs, forecast = NULL, year.target = NULL,
  detrend = FALSE, color.pal = c("bw", "reds", "ypb", "tcolor"),
  subtitle = NULL)
}
\arguments{
\item{hindcast}{A multi-member list with the hindcast for verification. See details.}

\item{obs}{List with the benchmarking observations for forecast verification.}

\item{forecast}{A multi-member list with the forecasts. Default is NULL.}

\item{year.target}{Year within the hindcast period considered as forecast. Default is NULL.}

\item{detrend}{Logical indicating if the data should be detrended. Default is FALSE.}

\item{color.pal}{Color palette for the representation of the probabilities. Default to \code{"bw"} (black and white).
\code{"reds"} for a white-red transition, \code{"ypb"} for a yellow-pink-blue transition  or
\code{"tcolor"} for a colorbar for each tercile, blue-grey-red for below, normal and above terciles, respectively.}

\item{subtitle}{String to include a subtitle bellow the title. Default is NULL.}
}
\description{
Tercile plot for the visualization of forecast skill of ensemble predictions.
 This function is prepared to plot the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
First daily data are averaged to obtain a single seasonal value. For
rectangular spatial domains (i.e., for fields), the spatial average is first computed (with a warning) to obtain a
unique series for the whole domain. The corresponding terciles for each ensemble member are then computed
for the hindcast period. Thus, each particular member and season, are categorized into three categories (above,
between or below), according to their respective climatological terciles. Then, a probabilistic forecast is computed
year by year by considering the number of members falling within each category. This probability is represented by the
colorbar. For instance, probabilities below 1/3 are very low, indicating that a minority of the members
falls in the tercile. Conversely, probabilities above 2/3 indicate a high level of member agreement (more than 66\% of members
falling in the same tercile). The observed terciles (the events that actually occurred) are represented by the
white circles. If the forecast object is not NULL, then the probabilities for this season are also ploted next to the
hindcast results.

Finally, the ROC Skill Score (ROCSS) is computed for the hindcasts. It is indicated in the secondary (right) Y axis.
For each tercile, it provides a quantitative measure of the forecast skill, and it is commonly used to evaluate the
performance of probabilistic systems (Joliffe and Stephenson 2003). The value of this score ranges from 1
(perfect forecast system) to -1 (perfectly bad forecast system). A value zero indicates no skill compared with a random
prediction. If year.target is not NULL, this year is not included in the computation of the score (operational point
of view).

In case of multi-member fields or stations, they are spatially averaged to obtain one single time series
for each member prior to data analysis, with a warning.
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\author{
M.D. Frias\email{mariadolores.frias@unican.es} and J. Fernandez based on the original diagram
conceived by A. Cofino (See Diez et al, 2011 for more details).
}
\references{
Diez, E., Orfila, B., Frias, M.D., Fernandez, J., Cofino, A.S., Gutierrez, J.M., 2011.
Downscaling ECMWF seasonal precipitation forecasts in Europe using the RCA model.
Tellus A 63, 757-762. doi:10.1111/j.1600-0870.2011.00523.x

Jolliffe, I. T. and Stephenson, D. B. 2003. Forecast Verification: A Practitioner's Guide in
Atmospheri Science, Wiley, NY
}

