% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bar.R
\name{theme_bar}
\alias{theme_bar}
\title{Custom Theme for Bar Charts}
\usage{
theme_bar(
  flip = TRUE,
  add_text = FALSE,
  axis_text_x_angle = 0,
  axis_text_x_vjust = 0.5,
  axis_text_x_hjust = 0.5
)
}
\value{
A custom theme object.
}
\description{
Custom Theme for Bar Charts
}
