% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_point.R
\name{theme_point}
\alias{theme_point}
\title{Custom Theme for Point Charts}
\usage{
theme_point()
}
\arguments{
\item{flip}{Logical. Whether the plot is flipped (horizonal).}

\item{axis_text_x_angle}{Angle for x-axis text.}

\item{axis_text_x_vjust}{Vertical justification for x-axis text.}

\item{axis_text_x_hjust}{Horizontal justification for x-axis text.}
}
\value{
A custom theme object.
}
\description{
Custom Theme for Point Charts
}
