% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reach.R
\name{theme_reach_borders}
\alias{theme_reach_borders}
\title{Some REACH theme for ggplot}
\usage{
theme_reach_borders(family = "Leelawadee")
}
\arguments{
\item{family}{The font family. Default to "Leelawadee"}
}
\value{
A theme to be added to the "+" ggplot grammar
}
\description{
Some REACH theme for ggplot
}
