% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reach.R
\name{theme_reach_flip_hist}
\alias{theme_reach_flip_hist}
\title{Some reach more minimal theme for ggplot}
\usage{
theme_reach_flip_hist(family = "Leelawadee")
}
\arguments{
\item{family}{The font family. Default to "Leelawadee"}
}
\value{
A theme to be added to the "+" ggplot grammar
}
\description{
Give some REACH colors and fonts to a ggplot. Based on theme_bw(). To be used for horizontal bar charts.
}
