% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_visualizer_bar.R
\docType{data}
\name{ThemeVisualizerBar}
\alias{ThemeVisualizerBar}
\title{ggplot2 theme for bar charts with sane defaults}
\format{
An object of class \code{ThemeVisualizerBar} (inherits from \code{ggproto}, \code{gg}) of length 1.
}
\usage{
ThemeVisualizerBar
}
\description{
ggplot2 theme for bar charts with sane defaults
}
\keyword{datasets}
