% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\name{tp.cfs.operative.2017}
\alias{tp.cfs.operative.2017}
\title{CFSv2 operative seasonal forecast of global DJF total precipitation amount (2017)}
\format{A grid}
\source{
Dowloaded from the ECOMS User Data Gateway \url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
CFSv2 operative seasonal forecast of DJF total precipitation amount for the entire globe of year 2017
 (i.e. December 2016, Januery-February 2017). Data are annually aggregated. Ensemble of 24 members,
  corresponding to the lead month 1 predictions (November 2016 initializations).
}
\examples{
\dontrun{
data(tp.cfs.operative.2017)
require(transformeR)
plotClimatology(climatology(tp.cfs.operative.2017, parallel = TRUE), backdrop.theme = "coastline")
}
}
\references{
Saha, S. et al. 2013. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}

