% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{tp.ncep}
\alias{tp.ncep}
\title{NCEP/NCAR reanalysis DJF total precipitation amount for the entire globe.}
\format{A grid}
\source{
Dowloaded from the ECOMS User Data Gateway \url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
NCEP/NCAR reanalysis DJF total precipitation amount for the entire globe (1949-2010). Data are annually aggregated.
}
\examples{
data(tp.ncep)
require(transformeR)
plotClimatology(climatology(tp.ncep), backdrop.theme = "coastline")
}

