% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_add_layer.R
\name{ui_add_layer}
\alias{ui_add_layer}
\title{ui_add_layer}
\usage{
ui_add_layer(btn_id, input, output, session)
}
\arguments{
\item{btn_id}{Id of the button clicked (type of layer to add)}

\item{input, }{output, session Standard Shiny parameters}
}
\value{
none
}
\description{
Add a Layer to the plot
}
\keyword{internal}
