% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_df_selection.R
\name{ui_df_selection}
\alias{ui_df_selection}
\title{ui_df_selection}
\usage{
ui_df_selection(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny parameters.}
}
\value{
name of the selected dataframe.
}
\description{
This function handles the display of a modal dialog to select a dataframe
from the Global Environment and returns the name of the selected dataframe.
}
\keyword{internal}
