% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_layer_events.R
\name{ui_layer_events}
\alias{ui_layer_events}
\title{ui_layer_events}
\usage{
ui_layer_events(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny parameters.}
}
\description{
Encapsulates the logic to handle the layer bar interaction.
}
\keyword{internal}
