% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_sidebar_layer.R
\name{ui_sidebar_layer}
\alias{ui_sidebar_layer}
\title{ui_sidebar_layer}
\usage{
ui_sidebar_layer()
}
\value{
Shiny UI Object
}
\description{
UI for the left sidebar used to select layer
}
\keyword{internal}
