% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_sidebar_parameters.R
\name{ui_sidebar_parameters}
\alias{ui_sidebar_parameters}
\title{ui_sidebar_parameters}
\usage{
ui_sidebar_parameters()
}
\value{
Shiny UI Object
}
\description{
UI for the left sidebar used to control parameters
}
\keyword{internal}
