% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useCSS.R
\name{useCSS}
\alias{useCSS}
\title{Import a CSS file either the package is installed or not}
\usage{
useCSS(file)
}
\arguments{
\item{file}{Name of the CSS file}
}
\description{
Import a CSS file either the package is installed or not
}
\keyword{internal}
