% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{vector2matrix}
\alias{vector2matrix}
\title{internal function concatenateDataRelDiagram_v2}
\usage{
vector2matrix(data, coordinates)
}
\arguments{
\item{data}{1*m (m = grid points) vector of data}

\item{coordinates}{m*2 (m = grid points) 2D-matrix of grid points. First (second) column corresponds to longitudes (latitudes)}
}
\value{
A list with three elements:
data: mx*my (mx = longitudes, my = latitudes) 2D-matrix of data
xnod: vector (length mx) of longitudes
ynod: vector (length my) of latitudes
}
\description{
Function to convert a vector of data into a matrix
}
\author{
R. Manzanas \& M.Iturbide
}
\keyword{internal}

