% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlot}
\alias{violinPlot}
\title{Lattice plot methods for climatological series}
\usage{
violinPlot(..., color.fun = list(FUN = sd, na.rm = TRUE),
  color.theme = "RdYlBu", color.cuts = NULL, rev.colors = FALSE,
  h.lines = NULL, lonLim = NULL, latLim = NULL,
  bwplot.custom = list())
}
\arguments{
\item{...}{Input grids (or station data)}

\item{color.fun}{list containing the function and the related arguments to perform spatial 
aggregation. The resulting values are used to fill with color the violins.
 Default is \code{list(FUN = sd, na.rm = TRUE)}.}

\item{color.theme}{A character string indicating the color theme to use in the map. 
Valid values are those available in the \code{\link{RColorBrewer}} themes. Additionally,
the \code{"jet.colors"} palette can be used (the rainbow colors, in general not advised, though),
for backwards compatibility. Default to the diverging, colorblind-friendly \code{"RdYlBu"} palette.}

\item{color.cuts}{Numeric sequence indicating the color cuts.}

\item{h.lines}{Numeric sequence indicating the position of dashed horizontal lines.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, 
in decimal degrees, of the bounding box selected. If NULL (default), the whole 
longitudinal range is selected}

\item{latLim}{Same as lonLim, but for the selection of the latitudinal range.}

\item{bwplot.custom}{List of arguments as passed to \link{\code{bwplot}}. Argument \code{panel} cannot 
be modified, thus, if specified, it will be ignored.}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
A wrapper for the lattice (trellis) plot methods for grid and station data.
}
\details{
The function applies the \code{\link[lattice]{xyplot}} method after computing spatial aggregation 
(parameter \code{aggr.spatial}) and member aggregation (mean and range) to the imput grids (or station data).
 
In case of multimember grids, the function will internally compute the ensemble mean 
and the range for plotting. The range is used to plot the shadow of the multimember spread.
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{bwplot}. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link[lattice]{trellis.par.set}}.
}
\examples{
data("EOBS_Iberia_tas")
data("EOBS_Iberia_pr")
data("CORDEX_Iberia_pr")
violinPlot("one" = climatology(EOBS_Iberia_pr), 
           "two" = climatology(EOBS_Iberia_tas), 
           "three" = climatology(CORDEX_Iberia_pr),
            h.lines = seq(0, 15, 5),
            color.cuts = seq(0, 3, 0.2),
            bwplot.custom = list(ylim = c(0, 20),
                                 ylab = "pr and tas"))
}
\author{
M. Iturbide
}
