% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassVisualization.R
\docType{class}
\name{visualization-class}
\alias{visualization-class}
\title{An S4 class that contains a visualization}
\value{
a visualization object
}
\description{
An S4 class that contains a visualization
}
\section{Slots}{

\describe{
\item{\code{baseVisuURL}}{baseURL to the github/npellet/visualizer index.html file}
}}

\examples{

new("visualization")

}
