% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeR-package.R
\docType{package}
\name{visualizeR-package}
\alias{visualizeR}
\alias{visualizeR-package}
\title{visualizeR: What a color! What a viz!}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It basically provides colors as hex codes, color palettes, and some viz functions (graphs and maps).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gnoblet/visualizeR}
  \item \url{https://gnoblet.github.io/visualizeR/}
}

}
\author{
\strong{Maintainer}: Noblet Guillaume \email{gnoblet@zaclys.net}

}
\keyword{internal}
