% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendData.R
\name{appendData}
\alias{appendData}
\title{This function allows to create, append data to be passed to the https://github.com/npellet/visualizer.
Prior to this, the list have to be parsed and stored as a json file. This file can be either stored locally
or stored on https://gist.github.com.}
\usage{
appendData(data, variable, variableName, type)
}
\arguments{
\item{data}{data, if it exists}

\item{variable}{new dataset}

\item{variableName}{name of the dataset. The name of the data must be exposed to the https://github.com/npellet/visualizer
using the exposeData.R function.}

\item{type}{type of the data. The types currently supported are, "1D Y array", "chart", "multiChart".}
}
\value{
a list object with the data
}
\description{
This function allows to create, append data to be passed to the https://github.com/npellet/visualizer.
Prior to this, the list have to be parsed and stored as a json file. This file can be either stored locally
or stored on https://gist.github.com.
}
\examples{
visu()
visu("https://my.cheminfo.org", "http://localhost:9898/view.json", "http://localhost:9898/data.json")
visu("https://my.cheminfo.org",
"https://gist.githubusercontent.com/jwist/57512dc74595a7c2b66c3e2bf4e76ba1/raw/test.view.json",
"https://gist.githubusercontent.com/jwist/3354274a2b1762e408af22a8217384d9/raw/data.json")
}
\seealso{
exposeData.R
}
