% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR6continents}
\alias{AR6continents}
\title{AR6 regions organized by continents from the IPCC 6th Assessment Report.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the IPCC 6th Assessment Report.
}
\examples{
\dontrun{
data(AR6continents)
sp::plot(AR6continents, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 World regions")
}
}
