% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{AR6regions}
\alias{AR6regions}
\title{AR6 regions from the IPCC 6th Assessment Report.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the IPCC 6th Assessment Report.
}
\details{
Region names are: "ALA", "NEC", "GIC", "NWN", "SWN", "CNA", "ENA", "CAM", "SSA", "NWS", 
"SAM", "SWS", "SES", "AMZ", "NEB", "MED", "SAH", "WAF", "EAF", "SWAF", "SEAF", "CAF", "NEA", "NWA", 
"WAS", "CAS", "TIB", "EAS", "SAS", "SEA", "NAU", "SAU", "NEU", "CEU", "ANT*", "ARC*", "NTP*", 
"STP*", "ETP*", "WIO*" and "CAR*".
}
\examples{
\dontrun{
data(AR6regions)

# All regions
sp::plot(AR6regions, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 World regions")

# A single region
sp::plot(AR6regions["MED"], border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("AR6 Mediterranean region")
}
}
