\name{balCol}
\alias{balCol}
\title{Baloise Colours}
\usage{
balCol(colour)
}
\description{
\code{balCol} returns the 6 different official colour in use at Baloise.
This saves some time!
}
\details{
select one of the following colours:
\itemize{
 \item{"dbl"}{ Dunkelblau}
 \item{"hbl"}{ Hellblau}
 \item{"vio"}{ Violet}
 \item{"rot"}{ Rot}
 \item{"ora"}{ Orange}
 \item{"gel"}{ Gelb}
}
}
\examples{
balCol("dbl")
}

