% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlot.R
\name{bubblePlot}
\alias{bubblePlot}
\title{Bubble plot for visualization of forecast skill of ensemble predictions.}
\usage{
bubblePlot(hindcast, obs, forecast = NULL, year.target = NULL,
  detrend = FALSE, score = TRUE, size.as.probability = TRUE,
  bubble.size = 1, score.range = NULL, piechart = FALSE,
  subtitle = NULL, color.reverse = FALSE, pch.neg.score = NULL,
  pch.obs.constant = NULL, pch.data.nan = NULL)
}
\arguments{
\item{hindcast}{A multi-member list with the hindcast for verification. See details.}

\item{obs}{List with the benchmarking observations for forecast verification.}

\item{forecast}{A multi-member list with the forecasts. Default is NULL.}

\item{year.target}{Year within the hindcast period considered as forecast. Default is NULL.}

\item{detrend}{Logical indicating if the data should be detrended. Default is FALSE.}

\item{score}{Logical indicating if the relative operating characteristic skill score (ROCSS) should be included. See 
details. Default is TRUE}

\item{size.as.probability}{Logical indicating if the tercile probabilities (magnitude proportional to bubble radius) 
are drawn in the plot. See details. Default is TRUE.}

\item{bubble.size}{Number for the bubble or pie size. bubble.size=1 by default.}

\item{score.range}{Two dim vector to rescale the transparency of the bubbles (e.g c(0.5, 0.8) values
lower than 0.5 are transparent while values up to 0.8 have maximum transparency). Default is NULL that is transparency 
ranges from 0 to 1.}

\item{piechart}{Logical flag indicating if pie charts should be plot instead of bubbles. Default is FALSE.}

\item{subtitle}{String to include a subtitle bellow the title. Default is NULL.}

\item{color.reverse}{Logical indicating if the color palete for the terciles (blue, grey, red) should be
reversed (e.g for precipitation). Default is FALSE.}

\item{pch.neg.score}{pch value to highlight the negative score values. Default is NULL. Not available for piecharts.}

\item{pch.obs.constant}{pch value to highlight those whose score cannot be computed due to constant obs 
conditions (e.g. always dry). Default is NULL.}

\item{pch.data.nan}{pch value to highlight those whose score cannot be computed due to time series with all NA values in 
the observations and/or models. If score=F, highlight those grids with all forecasts NaN.}
}
\description{
Bubble plot for visualization of forecast skill of ensemble predictions. It provides a
 spatially-explicit representation of the skill, resolution and reliability of a probabilistic predictive 
 system in a single map. 
 This function is prepared to plot the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See 
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
First daily data are averaged to obtain a single seasonal value. The corresponding terciles 
 for each ensemble member are then computed for the hindcast period. Thus, each particular grid point, member and season,
 are categorized into three categories (above, between or below), according to their respective climatological 
 terciles. Then, a probabilistic forecast is computed year by year by considering the number of members falling 
 within each category. For instance, probabilities below 1/3 are very low, indicating that a minority of the members 
 falls in the tercile. Conversely, probabilities above 2/3 indicate a high level of member agreement (more than 66\% of members
 falling in the same tercile). Probabilities are also computed for the forecast or the selected year. Color represents the tercile 
 with the highest probability for the forecast or selected year. The bubble size indicates the probability of that tercile. This 
 option is not plotted if the size.as.probability argument is FALSE.

 Finally, the ROC Skill Score (ROCSS) is computed for the hindcast period. For each tercile, it provides a quantitative measure 
 of the forecast skill, and it is commonly used to evaluate the performance of probabilistic systems (Joliffe and Stephenson 2003). 
 The value of this score ranges from 1 (perfect forecast system) to -1 (perfectly bad forecast system). A value zero indicates no 
 skill compared with a random prediction. The transparency of the bubble is associated to the ROCSS. By default only positive values
 are plotted if the score argument is TRUE. The target year is considered as forecast and it is not included in the computation of 
 the score (operational point of view).
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\references{
Jolliffe, I. T. and Stephenson, D. B. 2003. Forecast Verification: A Practitioner's Guide in Atmospheric 
 Science, Wiley, NY.
 
 Slingsby A., Lowe R., Dykes J., Stephenson D. B., Wood J., Jupp T. E. 2009. A pilot study for the collaborative 
 development of new ways of visualising seasonal climate forecasts. Proc. 17th Annu. Conf. of GIS Research UK, 
 Durham, UK, 1-3 April 2009.
}
\seealso{
Other VisualizeR: \code{\link{as.MrEnsemble}},
  \code{\link{as.MrGrid}}, \code{\link{as.MrStation}},
  \code{\link{convertIntoS4}}, \code{\link{spreadPlot}},
  \code{\link{tercileBarplot}}
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es} and J. Fernandez based on the original diagram 
 conceived by Slingsby et al 2009.
}
