% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\name{calcRank}
\alias{calcRank}
\title{Calculate a rank}
\usage{
calcRank(id, num_var, cat_var, data, decreasing = FALSE,
  ties.method = "random")
}
\arguments{
\item{id}{}

\item{num_var}{}

\item{cat_var}{}

\item{data}{}

\item{decreasing}{}

\item{ties.method}{}
}
\description{
\code{calcRank}
}

