% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{calculateReliability}
\alias{calculateReliability}
\title{Generate object needed by function reliability}
\usage{
calculateReliability(obs, prd, nbins = 3, nbinsprob = 10, nboot = 100,
  sigboot = 0.05)
}
\arguments{
\item{obs}{m*n matrix of observations (m = years, n = locations)}

\item{prd}{m*n*l matrix of predictions (m = members, n = years, l = locations)}

\item{nbins}{(optional): number of categories considered (e.g. 3 for terciles). By default nbins = 3}

\item{nbinsprob}{(optional): number of probability bins considered. By default nbinsprob = 10}

\item{nboot}{number of samples considered for bootstrapping. By default nboot = 100}

\item{sigboot}{sigboot}
}
\value{
List with the following elements:
nbins = nbins
nyear = number of years
npoint = number of locations
n = nyear*npoint
prdprob = probability bins (center), per category (e.g. per tercile)
obsfreq = observed frequency, per category (e.g. per tercile)
prdfreq = predicted frequency, per category (e.g. per tercile)
slope = slope of the reliability line, per category (e.g. per tercile)
slope_boot = nboot*nbins matrix, with all the boostrapped values for the slope of the reliability line
}
\description{
This function provides the object needed by "calculateReliability_v2.R" 
for calculating the reliability categories of a probabilistic prediction.
}
\author{
R. Manzanas \& M.Iturbide
}
\references{
Lawson, B.D. & Armitage, O.B., 2008. Weather guide for the Canadian Forest Fire Danger Rating System. Northern Forestry Centre, Edmonton (Canada).

van Wagner, C.E., 1987. Development and structure of the Canadian Forest Fire Weather Index (Forestry Tech. Rep. No. 35). Canadian Forestry Service, Ottawa, Canada.

van Wagner, C.E., Pickett, T.L., 1985. Equations and FORTRAN program for the Canadian forest fire weather index system (Forestry Tech. Rep. No. 33). Canadian Forestry Service, Ottawa, Canada.
}
\keyword{internal}

