% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{calculateReliability}
\alias{calculateReliability}
\title{Generate object needed by function reliability}
\usage{
calculateReliability(hindcast, obs, n.events = n.events, n.bins = n.bins,
  n.boot = n.boot)
}
\arguments{
\item{hindcast}{m*n*l matrix of predictions (m = members, n = years, l = locations)}

\item{obs}{m*n matrix of observations (m = years, n = locations)}

\item{n.events}{(optional): number of categories considered (e.g. 3 for terciles). By default n.events = 3}

\item{n.bins}{(optional): number of probability bins considered. By default n.bins = 10}

\item{n.boot}{number of samples considered for bootstrapping. By default n.boot = 100}
}
\value{
List with the following elements:
n.events = n.events
nyear = number of years
npoint = number of locations
n = nyear*npoint
hindcastprob = probability bins (center), per event (e.g. per tercile)
obsfreq = observed frequency, per event (e.g. per tercile)
hindcastfreq = predicted frequency, per event (e.g. per tercile)
slope = slope of the reliability line, per event (e.g. per tercile)
slope_boot = n.boot*n.events matrix, with all the boostrapped values for the slope of the reliability line
}
\description{
This function provides the object needed by "calculateReliability_v2.R"
for calculating the reliability categories of a probabilistic prediction.
}
\author{
R. Manzanas \& M.Iturbide
}
\keyword{internal}

