% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColor.R
\name{getColor}
\alias{getColor}
\title{GetColorChart based on character (not used)}
\usage{
getColor(text)
}
\arguments{
\item{text}{a text used to pick the color}
}
\value{
a rgba color string
}
\description{
GetColorChart based on character (not used)
}
\examples{
getColor("colombia")
}
