% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbar.R
\name{hbar_percent}
\alias{hbar_percent}
\title{Simple horizontal bar chart which scales x_labels to percentages}
\usage{
hbar_percent(
  .tbl,
  x,
  y,
  group = NULL,
  initiative = "reach",
  palette = "primary",
  width = 0.5,
  x_title = "",
  y_title = "",
  group_title = NULL,
  font_family = "Leelawadee",
  position = "dodge",
  reverse = FALSE,
  title = "",
  subtitle = "",
  gg_theme = NULL,
  ...
)
}
\arguments{
\item{.tbl}{Some data}

\item{x}{Some numeric column on the x scale}

\item{y}{Some column on the y scale}

\item{group}{Some grouping categorical column, e.g. administrative areas}

\item{initiative}{Either "reach" or "agora" or "impact" for the color palette}

\item{palette}{The color palette from the initiative}

\item{width}{Width}

\item{x_title}{The x scale title. Default to empty string}

\item{y_title}{The y scale title. Default to empty string}

\item{group_title}{The group legend title. Defaut to NULL}

\item{font_family}{The font family. Default to "Leelawadee"}

\item{position}{Should the chart be stacked? Default to dodge}

\item{reverse}{Boolean indicating whether the color palette should be reversed}

\item{title}{Plot title. Default to empty string}

\item{subtitle}{Plot subtitle. Default to empty string}

\item{gg_theme}{Some ggplot2 theme}

\item{...}{Other arguments to be passed to "ggblanket::gg_col"}
}
\value{
A horizontal bar chart
}
\description{
Simple horizontal bar chart which scales x_labels to percentages
}
