% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{lollipop}
\alias{lollipop}
\title{Simple bar chart}
\usage{
lollipop(
  df,
  x,
  y,
  flip = TRUE,
  wrap = NULL,
  arrange = TRUE,
  point_size = 3,
  point_color = cols_reach("main_red"),
  point_alpha = 1,
  segment_size = 1,
  segment_color = cols_reach("main_grey"),
  segment_alpha = 1,
  alpha = 1,
  x_title = NULL,
  y_title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  add_text = FALSE,
  add_text_size = 3,
  add_text_suffix = "",
  add_text_color = "white",
  add_text_fontface = "bold",
  theme = theme_reach()
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric column.}

\item{y}{A character column or coercible as a character column.}

\item{flip}{TRUE or FALSE. Default to TRUE or horizontal lollipop plot.}

\item{wrap}{Should x-labels be wrapped? Number of characters.}

\item{arrange}{TRUE or FALSE. Arrange by highest percentage first.}

\item{point_size}{Point size.}

\item{point_color}{Point color.}

\item{point_alpha}{Point alpha.}

\item{segment_size}{Segment size.}

\item{segment_color}{Segment color.}

\item{segment_alpha}{Segment alpha.}

\item{alpha}{Fill transparency.}

\item{x_title}{The x scale title. Default to NULL.}

\item{y_title}{The y scale title. Default to NULL.}

\item{title}{Plot title. Default to NULL.}

\item{subtitle}{Plot subtitle. Default to NULL.}

\item{caption}{Plot caption. Default to NULL.}

\item{add_text}{TRUE or FALSE. Add the y value as text within the bubble.}

\item{add_text_size}{Text size.}

\item{add_text_suffix}{If percent is FALSE, should we add a suffix to the text label?}

\item{add_text_color}{Added text color. Default to white.}

\item{add_text_fontface}{Added text font face. Default to "bold".}

\item{theme}{Whatever theme. Default to theme_reach().}
}
\value{
A bar chart
}
\description{
Simple bar chart
}
