% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeShape}
\alias{mergeShape}
\title{Load and prepare shape file}
\usage{
mergeShape(df, path_shape, df_ID, shape_ID)
}
\arguments{
\item{df}{Data that has to be merged with the shape file.}

\item{path_shape}{Path to the shape file.}

\item{df_ID}{Unique ID of the data file for the join.}

\item{shape_ID}{Unique ID of the shape file for the join.}
}
\description{
\code{mergeShape} combines a shape file with a designated dataframe
according to an ID variable.
}

