% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Simple point chart}
\usage{
point(
  df,
  x,
  y,
  group = "",
  add_color = color("branding_reach_red"),
  flip = TRUE,
  alpha = 1,
  size = 2,
  x_title = NULL,
  y_title = NULL,
  group_title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  theme_fun = theme_reach(grid_major_y = TRUE),
  palette = "cat_5_ibm",
  scale_impact = TRUE,
  direction = 1,
  reverse_guide = TRUE
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric column.}

\item{y}{Another numeric column.}

\item{group}{Some grouping categorical column, e.g. administrative areas or population groups.}

\item{add_color}{Add a color to bars (if no grouping).}

\item{flip}{TRUE or FALSE. Default to TRUE or horizontal bar plot.}

\item{alpha}{Fill transparency.}

\item{size}{Point size.}

\item{x_title}{The x scale title. Default to NULL.}

\item{y_title}{The y scale title. Default to NULL.}

\item{group_title}{The group legend title. Default to NULL.}

\item{title}{Plot title. Default to NULL.}

\item{subtitle}{Plot subtitle. Default to NULL.}

\item{caption}{Plot caption. Default to NULL.}

\item{theme_fun}{Whatever theme. Default to theme_reach(). NULL if no theming needed.}

\item{scale_impact}{Use the package custom scales for fill and color.}
}
\description{
Simple point chart
}
