% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Simple scatterplot}
\usage{
point(
  df,
  x,
  y,
  group = "",
  facet = "",
  facet_scales = "free",
  x_rm_na = TRUE,
  y_rm_na = TRUE,
  group_rm_na = TRUE,
  facet_rm_na = TRUE,
  add_color = color("cat_5_main_1"),
  add_color_guide = TRUE,
  flip = TRUE,
  alpha = 1,
  size = 2,
  x_title = NULL,
  y_title = NULL,
  group_title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  theme_fun = theme_point(),
  scale_fill_fun = scale_fill_visualizer_discrete(),
  scale_color_fun = scale_color_visualizer_discrete()
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A quoted numeric column.}

\item{y}{A quoted numeric column.}

\item{group}{Some quoted grouping categorical column, e.g. administrative areas or population groups.}

\item{facet}{Some quoted grouping categorical column.}

\item{facet_scales}{Character. Either "free" (default) or "fixed" for facet scales.}

\item{x_rm_na}{Remove NAs in x?}

\item{y_rm_na}{Remove NAs in y?}

\item{group_rm_na}{Remove NAs in group?}

\item{facet_rm_na}{Remove NAs in facet?}

\item{add_color}{Add a color to points (if no grouping).}

\item{add_color_guide}{Should a legend be added?}

\item{flip}{TRUE or FALSE.}

\item{alpha}{Fill transparency.}

\item{size}{Point size.}

\item{x_title}{The x scale title. Default to NULL.}

\item{y_title}{The y scale title. Default to NULL.}

\item{group_title}{The group legend title. Default to NULL.}

\item{title}{Plot title. Default to NULL.}

\item{subtitle}{Plot subtitle. Default to NULL.}

\item{caption}{Plot caption. Default to NULL.}

\item{theme_fun}{Whatever theme. Default to theme_point(). NULL if no theming needed.}

\item{scale_fill_fun}{Scale fill function. Default to scale_fill_visualizer_discrete().}

\item{scale_color_fun}{Scale color function. Default to scale_color_visualizer_discrete().}
}
\description{
Simple scatterplot
}
