% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{prd2prob}
\alias{prd2prob}
\title{internal function prob2bin}
\usage{
prd2prob(prd, nbins, prd4cats = NULL)
}
\arguments{
\item{prd}{3D-array of predictions, dimensions = (member, time, npoints)}

\item{nbins}{Number of categories (3 for terciles)}

\item{prd4cats}{Optional. 3D-array of predictions for which calculate the categories (e.g., terciles)}

\item{dimensions}{(member, time, npoints)}
}
\value{
prdprob: 3D-array of probabilistic predictions, dimensions = (nbins, time, npoints)
}
\description{
This function provides the object needed by "calculateReliability_v2.R" 
for calculating the reliability categories of a probabilistic prediction.
}
\note{
For prd4cats, categories are calculated at model- (not at member-) level
}
\author{
R. Manzanas \& M.Iturbide
}
\keyword{internal}
