% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\name{PRUDENCEregions}
\alias{PRUDENCEregions}
\title{PRUDENCE regions.}
\format{A SpatialPolygons object}
\description{
SpatialPolygons object of the geographical regions used in the PRUDENCE project.
}
\examples{
\dontrun{
data(PRUDENCEregions)
plot(PRUDENCEregions)
}
}

