% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\name{PRUDENCEregions}
\alias{PRUDENCEregions}
\title{PRUDENCE regions.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the PRUDENCE project.
}
\examples{
\dontrun{
data(PRUDENCEregions)
sp::plot(PRUDENCEregions, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("PRUDENCE European regions")
}
}

