% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{PRUDENCEregions}
\alias{PRUDENCEregions}
\title{PRUDENCE regions.}
\format{A \code{\link[sp]{SpatialPolygons-class}} object}
\description{
Spatial Polygons of the geographical regions used in the PRUDENCE project.
}
\details{
For further detail, see \url{http://ensemblesrt3.dmi.dk/quicklook/regions.html}
}
\examples{
\dontrun{
data(PRUDENCEregions)
sp::plot(PRUDENCEregions, border = "red", axes = TRUE)
transformeR:::draw.world.lines()
title("PRUDENCE European regions")
}
}
