% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{reliabilityCategories}
\alias{reliabilityCategories}
\title{Reliability categories of a probabilistic prd.}
\usage{
reliabilityCategories(obs, prd, regions = NULL, nbins = 3, nbinsprob = 10,
  nboot = 100, sigboot = 0.05, diagrams = TRUE)
}
\arguments{
\item{obs}{Grid of observations}

\item{prd}{Grid of forecast data}

\item{regions}{SpatialPolygons* object \code{\link[sp]{SpatialPolygons}}. delimiting the regions for which
the relaiability is calculated. Default is NULL (See details).}

\item{nbins}{(optional): number of categories considered (e.g. 3 for terciles). By default nbins = 3}

\item{nbinsprob}{(optional): number of probability bins considered. By default nbinsprob = 10}

\item{nboot}{number of samples considered for bootstrapping. By default nboot = 100}

\item{sigboot}{Optional. Confidence interval for the reliability line. By default sigboot = 0.05}

\item{diagrams}{Logical (default = TRUE). Plotting results.}
}
\value{
Grid of reliability categories with an additional data dimension for categories
and an additional slot ($ReliabilityCategories) containing the following elements:
catname: reliability categories.
slope: \code{$slope} slope of the reliability line; \code{$lower} lower bound confidence for slope (according to "sigboot");
\code{$upper} upper bound confidence for slope (according to "sigboot").
}
\description{
Calculates (and draws)
reliability diagrams and the related reliability categories, according to Weisheimer et al. 2014.
}
\details{
If parameter regions is NULL (default) the whole region in \code{obs} and \code{prd} is considered
for computing reliability. A subregion will be considered If the corresponding SpatialPolygons* object
is provided. In these cases, if parameter \code{diagrams = TRUE}, reliability diagrams are plotted for
each specified category. If a SpatialPolygons* object of multiple subregions is provided, reliability is computed
separately for each region and reliability maps are plotted instead.

A SpartialPolygons* object is easily obtained by reading a shapefile with function
\code{\link[rgdal]{readOGR}}.
}
\examples{
\dontrun{
data("tas.cfs")
data("tas.ncep")
data("AR5regions")
require(transformeR)
tas.cfs2 <- interpGrid(tas.cfs, getGrid(tas.ncep))
tas.ncep2 <- subsetGrid(tas.ncep, years = 1983:2009)
tas.cfs2 <- subsetGrid(tas.cfs2, years = 1983:2009)
rel.reg <- reliabilityCategories(obs = tas.ncep2, prd = tas.cfs2,
                           nboot = 100, regions = AR5regions)
rel <- reliabilityCategories(obs = tas.ncep2, prd = tas.cfs2,
                           nboot = 100)
}
}
\author{
R. Manzanas \& M.Iturbide
}
\references{
Weisheimer, A., Palmer, T.N., 2014. On the reliability of seasonal climate forecasts. Journal of The Royal Society Interface 11, 20131162. doi:10.1098/rsif.2013.1162
}

