% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliabilityCategories.R
\name{reliabilityCategories}
\alias{reliabilityCategories}
\title{Reliability categories of a probabilistic prediction.}
\usage{
reliabilityCategories(obs, prd, nbins = 3, nbinsprob = 10, nboot = 100,
  sigboot = 0.05, diagrams = TRUE)
}
\arguments{
\item{obs}{Grid of observations}

\item{prd}{Grid of predictions}

\item{nbins}{(optional): number of categories considered (e.g. 3 for terciles). By default nbins = 3}

\item{nbinsprob}{(optional): number of probability bins considered. By default nbinsprob = 10}

\item{nboot}{number of samples considered for bootstrapping. By default nboot = 100}

\item{sigboot}{Optional. Confidence interval for the reliability line. By default sigboot = 0.05}

\item{diagrams}{Logical (default = TRUE). Plotting results.}

\item{nod}{Required if diagrams = TRUE. m*2 matrix of coordinates (m=locations, column1=latitude, column2=longitude)}

\item{xlim}{Required if diagrams = TRUE. Limits for maps}

\item{ylim}{Required if diagrams = TRUE. Limits for maps}
}
\value{
Same as prd but with an aditional list ($ReliabilityCategories) containing the following elements:
catcol: color of the reliability category
catname: reliability category
sl: slope of the reliability line
sl_lower: lower bound confidence for slope (according to "sigboot")
sl_lower: upper bound confidence for slope (according to "sigboot")
}
\description{
This function calculates (and draws)
reliability diagrams and the related reliability categories, according to Weisheimer et al. 2014:
http://rsif.royalsocietypublishing.org/content/11/96/20131162
}
\author{
R. Manzanas \& M.Iturbide
}

