% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.col.R
\name{rep.col}
\alias{rep.col}
\title{repeat a column when needed to match other matrix dimension}
\usage{
\method{rep}{col}(c, n)
}
\arguments{
\item{c}{a vector to be repeated}

\item{n}{number ot time to repeat the vector}
}
\value{
a matrix
}
\description{
repeat a column when needed to match other matrix dimension
}
\references{
the original function was found here: \url{https://www.r-bloggers.com/a-quick-way-to-do-row-repeat-and-col-repeat-rep-row-rep-col/}
}
