% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_color}
\alias{scale_color}
\title{Color scale constructor for REACH or AGORA colors}
\usage{
scale_color(
  initiative = "reach",
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  reverse_guide = TRUE,
  ...
)
}
\arguments{
\item{initiative}{Either "reach" or "agora" or "default".}

\item{palette}{Palette name from `pal_reach()` or `pal_agora()`.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{reverse_guide}{Boolean indicating whether the guide should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradient(), used respectively when discrete is TRUE or FALSE.}
}
\value{
A color scale for ggplot
}
\description{
Color scale constructor for REACH or AGORA colors
}
