% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spreadPlot.R
\name{spreadPlot}
\alias{spreadPlot}
\title{Box plot for visualization of forecast skill of ensemble predictions.}
\usage{
spreadPlot(mm.obj, year.target, detrend = FALSE, boxplot = TRUE,
  violin = FALSE, add.points = FALSE, pch = NULL)
}
\arguments{
\item{mm.obj}{A multi-member list with predictions, either a field or a multi-member
station object as a result of downscaling of a forecast using station data. See details.}

\item{year.target}{Year selected to plot the probability of the tercile in bars.}

\item{detrend}{Logical indicating if the data should be detrended. Default is TRUE.}

\item{boxplot}{Logical flag indicating whether a boxplot should be added to the graph. Default is TRUE.}

\item{violin}{Logical flag indicating whether a violin plot should be added to the graph instead of the boxplot.
Default is FALSE.}

\item{add.points}{Logical flag indicating whether crosses indicating the ensemble members should be added to the
graph. Default is FALSE.}

\item{pch}{Single character or an integer code for one of the set of R graphics symbols. Cross is the default symbol.
 Only consider if add.points is TRUE.}
}
\description{
Box plot for visualization of forecast skill of ensemble predictions.
 This function is prepared to plot the data sets loaded from the ECOMS User Data Gateway (ECOMS-UDG). See
 the loadeR.ECOMS R package for more details (http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage).
}
\details{
For rectangular spatial domains (i.e., for fields), the spatial average is first computed (with a warning) to obtain a
 unique series for the whole domain. The climatology for the forecast period is computed. The shaded areas show
 the central tercile (dark shade) and the maximum and minimum (light shade). To avoid overinterpretation of daily
 peaks, the daily data has been smoothed by means of a (centered) moving average of 31 days. Therefore, at the
 location of the boxplots, the background shows the monthly mean forecast (the terciles and extremes being computed
 over members and years). Monthly means are calculated for each ensemble member. This information is included to the
 graph in a boxplot or a violin plot (to unveil multimodalities in the data) or/and the values of the ensemble members.

 In case of multi-member fields or stations, they are spatially averaged to obtain one single time series
 for each member prior to data analysis, with a warning.
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\author{
M.D. Frias \email{mariadolores.frias@unican.es} and J. Fernandez
}
\seealso{
Other VisualizeR: \code{\link{as.MrEnsemble}};
  \code{\link{as.MrGrid}}; \code{\link{as.MrStation}};
  \code{\link{bubblePlot}}; \code{\link{convertIntoS4}};
  \code{\link{tercileBarplot}}
}

