% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{tas.cfs}
\alias{tas.cfs}
\title{NCEP CFSv2 hindcast of DJF mean temperature for the entire globe.}
\format{
A grid
}
\source{
Dowloaded from the ECOMS User Data Gateway \url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
NCEP CFSv2 hindcast of December to February (DJF) mean temperature for the entire globe (1983-2010). Data are annually aggregated.
Ensemble of 24 members, corresponding to the lead month 1 predictions (November initializations)

For further detail on ensemble member definition, see \url{http://www.meteo.unican.es/ecoms-udg/dataserver/datasets/CFSv2}
}
\examples{
\dontrun{
data(tas.cfs)
spatialPlot(climatology(tas.cfs), backdrop.theme = "coastline", rev.colors = TRUE)
}
}
\references{
Saha, S. et al. 2014. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}
