% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reach.R
\name{theme_reach}
\alias{theme_reach}
\title{Base REACH ggplot2 theme}
\usage{
theme_reach(family = "Leelawadee")
}
\arguments{
\item{family}{The font family. Default to "Leelawadee"}
}
\value{
The base REACH theme
}
\description{
Give some reach colors and fonts to a ggplot. Based on theme_bw()
}
