% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlot}
\alias{violinPlot}
\title{Lattice plot methods for climatological series}
\usage{
violinPlot(..., group.index = NULL, color.fun = list(FUN = mean, na.rm
  = TRUE), color.theme = "RdYlBu", color.cuts = NULL,
  rev.colors = FALSE, h.lines = NULL, lonLim = NULL, latLim = NULL,
  bwplot.custom = list())
}
\arguments{
\item{...}{Input grids (or station data)}

\item{group.index}{Character or numeric passed to argument group in bwplot.}

\item{color.fun}{list containing the function and the related arguments to perform spatial 
aggregation. The resulting values are used to fill with color the violins.
 Default is \code{list(FUN = mean, na.rm = TRUE)}.}

\item{color.theme}{A character string indicating the color theme to use in the map. 
Valid values are those available in the \code{\link{RColorBrewer}} themes. Additionally,
the \code{"jet.colors"} palette can be used (the rainbow colors, in general not advised, though),
for backwards compatibility. Default to the diverging, colorblind-friendly \code{"RdYlBu"} palette.}

\item{color.cuts}{Numeric sequence indicating the color cuts.}

\item{rev.colors}{Default to FALSE. If TRUE the reversed version of the color palette
is used.}

\item{h.lines}{Numeric sequence indicating the position of dashed horizontal lines.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, 
in decimal degrees, of the bounding box selected. If NULL (default), the whole 
longitudinal range is selected}

\item{latLim}{Same as lonLim, but for the selection of the latitudinal range.}

\item{bwplot.custom}{List of arguments as passed to \code{\link[lattice]{bwplot}}. Arguments \code{panel} cannot 
be modified, thus, if specified, it will be ignored. 
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the plot can be controlled passing the relevant arguments to 
 \code{bwplot}. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link[lattice]{trellis.par.set}}.}
}
\value{
A lattice plot of class \dQuote{trellis}.
}
\description{
A wrapper for the lattice (trellis) plot methods for grid and station data.
}
\examples{
data("EOBS_Iberia_tas")
data("EOBS_Iberia_pr")
data("CORDEX_Iberia_pr")
violinPlot("one" = climatology(EOBS_Iberia_pr), 
           "two" = climatology(EOBS_Iberia_tas), 
           "three" = climatology(CORDEX_Iberia_pr),
            h.lines = seq(0, 15, 5),
            color.cuts = seq(0, 3, 0.2),
            bwplot.custom = list(ylim = c(0, 20),
                                 ylab = "pr and tas"))
                                 
## With grouping:
data("EOBS_Iberia_tas")
data("EOBS_Iberia_pr")
data("CORDEX_Iberia_pr")
data("CORDEX_Iberia_tas")
violinPlot("pr" = climatology(EOBS_Iberia_pr),
           "tas" = climatology(EOBS_Iberia_tas),
           "pr" = climatology(CORDEX_Iberia_pr),
           "tas" = climatology(CORDEX_Iberia_tas),
            group.index = c("Measure1", "Measure1", "Measure2", "Measure2"),
            h.lines = seq(0, 15, 5),
            color.cuts = seq(0, 3, 0.2),
            bwplot.custom = list(ylim = c(0, 20),
                                 ylab = "pr and tas",
                                 as.table = TRUE))
}
\author{
M. Iturbide
}
