% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{visualizeR-package}
\alias{visualizeR}
\alias{visualizeR-package}
\title{Generate R documentation from inline comments.}
\description{
VisualizeR is for the one
}
\details{
The only function you're likely to need from \pkg{visualizeR} is
\code{\link{visualizeR}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
